Add components folder
