+++ client/components/Admin/index.js
@@ -1,0 +1,674 @@
+import React from "react";
+
+const Admin = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={100}
+    height={100}
+    viewBox="0 0 646 646"
+  >
+    <style jsx>
+      {`
+        .adminst0 {
+          fill: #273942;
+        }
+        .adminst1 {
+          fill: #b7cdd8;
+        }
+        .adminst2 {
+          fill: #e2e2f7;
+        }
+        .adminst3 {
+          fill: #52677a;
+        }
+        .adminst4 {
+          fill: #354f5c;
+        }
+        .adminst5 {
+          fill: #ce521f;
+        }
+        .adminst6 {
+          fill: #ebebeb;
+        }
+        .adminst7 {
+          fill: #8e96a3;
+        }
+        .adminst8 {
+          fill: #3d4d5c;
+        }
+        .adminst9 {
+          fill: #3e9697;
+        }
+        .adminst10 {
+          fill: #68b7ce;
+        }
+        .adminst11 {
+          fill: #e8e8e8;
+        }
+        .adminst12 {
+          fill: #ffffff;
+        }
+        .adminst13 {
+          fill: #828282;
+        }
+        .adminst14 {
+          fill: #cccccc;
+        }
+        .adminst15 {
+          fill: #606060;
+        }
+        .adminst16 {
+          fill: #c6c6c6;
+        }
+        .adminst17 {
+          opacity: 0.15;
+        }
+        .adminst18 {
+          fill: #1d1e1c;
+        }
+        .adminst19 {
+          fill: #ffd700;
+        }
+        .adminst20 {
+          fill: #ff9700;
+        }
+        .adminst21 {
+          fill: #ffffe8;
+        }
+        .adminst22 {
+          fill: #f2dab8;
+        }
+        .adminst23 {
+          fill: #d44d41;
+        }
+        .adminst24 {
+          opacity: 0.3;
+        }
+        .adminst25 {
+          opacity: 0.3;
+          fill: #ffffff;
+        }
+        .adminst26 {
+          fill: #f01c01;
+        }
+        .adminst27 {
+          fill: #dbdad9;
+        }
+        .adminst28 {
+          opacity: 0.4;
+          fill: #ffffff;
+        }
+        .adminst29 {
+          opacity: 0.2;
+        }
+        .adminst30 {
+          opacity: 0.2;
+          fill: #020202;
+        }
+        .adminst31 {
+          fill: #f3554b;
+        }
+        .adminst32 {
+          fill: #bf2c28;
+        }
+        .adminst33 {
+          fill: #1f6a7b;
+        }
+        .adminst34 {
+          fill: #0594af;
+        }
+        .adminst35 {
+          fill: #38a9b4;
+        }
+        .adminst36 {
+          opacity: 0.2;
+          fill: #1d1e1c;
+        }
+        .adminst37 {
+          opacity: 0.64;
+          fill: #a6d9ed;
+        }
+        .adminst38 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #f0f0f0;
+        }
+        .adminst39 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #e1dddd;
+        }
+        .adminst40 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #ffffff;
+        }
+        .adminst41 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #af3737;
+        }
+        .adminst42 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #990000;
+        }
+        .adminst43 {
+          fill: #990000;
+        }
+        .adminst44 {
+          fill: #510000;
+        }
+        .adminst45 {
+          opacity: 0.5;
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #8e2525;
+        }
+        .adminst46 {
+          fill: #2ba2b2;
+        }
+        .adminst47 {
+          fill: #34aac0;
+        }
+        .adminst48 {
+          fill: #f69f00;
+        }
+        .adminst49 {
+          fill: #f59e00;
+        }
+        .adminst50 {
+          opacity: 0.1;
+        }
+        .adminst51 {
+          fill: #ed9406;
+        }
+        .adminst52 {
+          fill: #c6006d;
+        }
+        .adminst53 {
+          fill: #a50263;
+        }
+        .adminst54 {
+          fill: #6ac700;
+        }
+        .adminst55 {
+          fill: #5b9904;
+        }
+        .adminst56 {
+          fill: #070707;
+        }
+      `}
+    </style>
+    <path
+      className="adminst0"
+      d="M77.9 459.8c0 0-1.2 11.5 33.3 11.5 34.5 0 438.5 0.7 438.5 0.7s23.1 0.7 29.3-11.3L77.9 459.8z"
+    />
+    <path
+      className="adminst1"
+      d="M579 457.5c-2 4.3-7.8 6.5-13.8 6.5l-473.7-0.7c-6 0-12.1-3-13.7-6.7l33.2-66.2c3.8-5.8 3.3-8.5 10.2-8.5h416.2c3.3 0 7.6 2.4 10.5 8.5L579 457.5z"
+    />
+    <path
+      className="adminst2"
+      d="M335.3 140.7c0 1.6-1.3 2.9-2.8 2.9h-5.7c-1.6 0-2.8-1.3-2.8-2.9l0 0c0-1.6 1.3-2.9 2.8-2.9h5.7C334.1 137.8 335.3 139.1 335.3 140.7L335.3 140.7z"
+    />
+    <polygon
+      className="adminst3"
+      points="255.9 435.6 259.2 424.8 232.5 424.8 227.9 435.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="262 421.1 265 411.8 239.2 411.8 235.1 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="265.8 408.6 268.8 399.2 243.9 399.2 239.9 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="228.4 408.6 232.2 399.2 209.1 399.2 204.3 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="216.1 435.6 220.5 424.8 195.7 424.8 190.1 435.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="223 421.1 226.9 411.8 202.9 411.8 198.1 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="298.8 408.6 301 399.2 279.9 399.2 276.8 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="193.2 408.6 197.9 399.2 173.1 399.2 167.5 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="184.5 424.8 158.8 424.8 151.4 435.6 179.2 435.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="186.3 421.3 191.6 412 166.5 412 160.7 421.3 "
+    />
+    <polygon
+      className="adminst3"
+      points="297 421.1 299 411.8 274.9 411.8 272.7 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="155.3 412 132.9 412 124.6 420.6 149.4 420.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="146 424.5 122.7 424.5 113.4 434.9 139 434.9 "
+    />
+    <polygon
+      className="adminst3"
+      points="157.9 408 163.2 399.2 141.7 399.2 135.4 408 "
+    />
+    <polygon
+      className="adminst3"
+      points="474.1 435.6 501.9 435.6 494.7 424.8 468.1 424.8 "
+    />
+    <polygon
+      className="adminst3"
+      points="460.7 411.8 465.9 421.1 492.7 421.1 486.3 411.8 "
+    />
+    <polygon
+      className="adminst3"
+      points="485.6 408.6 479.4 399.2 454.7 399.2 459.9 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="413.1 421.1 408.5 411.8 385.7 411.8 389.4 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="419.9 435.6 413.9 424.8 390.4 424.8 394.6 435.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="445.7 408.6 441.7 399.2 417.6 399.2 422.8 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="453.7 421.1 448.2 411.8 423.9 411.8 428.4 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="434.3 435.6 460.6 435.6 454.3 424.8 429.2 424.8 "
+    />
+    <polygon
+      className="adminst3"
+      points="408.1 408.6 402.2 399.2 378.8 399.2 383.7 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="509.3 424.8 516.2 435.6 545.5 435.6 537.3 424.8 "
+    />
+    <polygon
+      className="adminst3"
+      points="526.7 408.6 519.5 399.2 493.5 399.2 499.5 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="507.2 421.1 535.3 421.1 528.1 411.8 501.1 411.8 "
+    />
+    <polygon
+      className="adminst3"
+      points="372.5 408.6 368.6 399.2 343.9 399.2 346.7 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="376.1 421.1 372.2 411.8 346.5 411.8 349.4 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="334.9 408.6 333.9 399.2 311 399.2 310.1 408.6 "
+    />
+    <polygon
+      className="adminst3"
+      points="336 421.1 334.4 411.8 310.1 411.8 308.9 421.1 "
+    />
+    <polygon
+      className="adminst3"
+      points="376.9 424.8 349.2 424.8 324.2 424.8 324 424.8 298.2 424.8 297.8 424.8 270.6 424.8 268.1 435.6 297 435.6 297.4 435.6 324.5 435.6 324.6 435.6 350.8 435.6 380.4 435.6 "
+    />
+    <path
+      className="adminst0"
+      d="M269.9 453.7l-1.4 0.7c-4.2 3.9-0.5 5 10.4 5h86.6c10.9 0 14.9-1.5 10.2-5l-1.4-0.7H269.9z"
+    />
+    <path
+      className="adminst4"
+      d="M364.5 446.7c-2.9-1.7-4.9-2.6-11.1-2.6l-63-0.1c-6.2 0-7.8 1-10.7 2.6l-9.8 6.9h104.3L364.5 446.7z"
+    />
+    <path
+      className="adminst5"
+      d="M122.9 274.2L122.9 274.2c-0.8-0.3-1.2-1.2-0.9-2l1.3-3.4 2.9 1.1 -1.3 3.4C124.6 274 123.7 274.5 122.9 274.2z"
+    />
+    <path
+      className="adminst5"
+      d="M139.2 280.2L139.2 280.2c-0.8-0.3-1.2-1.2-0.9-2 0 0 1.2-3.2 1.3-3.4 0.1-0.2 2.9 1.1 2.9 1.1l-1.3 3.4C140.9 280 140 280.5 139.2 280.2z"
+    />
+    <path
+      className="adminst6"
+      d="M122.6 273.5L122.6 273.5c-0.8-0.3-1.2-1.2-0.9-2l1.4-3.8c0.3-0.8 1.2-1.2 2-0.9l0 0c0.8 0.3 1.2 1.2 0.9 2l-1.4 3.8C124.3 273.4 123.4 273.8 122.6 273.5z"
+    />
+    <path
+      className="adminst6"
+      d="M138.8 279.5L138.8 279.5c-0.8-0.3-1.2-1.2-0.9-2l1.4-3.8c0.3-0.8 1.2-1.2 2-0.9v0c0.8 0.3 1.2 1.2 0.9 2l-1.4 3.8C140.5 279.3 139.6 279.8 138.8 279.5z"
+    />
+    <rect
+      x="117.2"
+      y={311}
+      transform="matrix(0.9384 0.3456 -0.3456 0.9384 116.1504 -22.8967)"
+      className="adminst7"
+      width="10.2"
+      height="6.6"
+    />
+    <rect
+      x="121.6"
+      y={300}
+      transform="matrix(0.9384 0.3456 -0.3456 0.9384 112.6343 -25.0805)"
+      className="adminst7"
+      width="10.2"
+      height="6.6"
+    />
+    <rect
+      x="114.2"
+      y="102.3"
+      className="adminst8"
+      width="424.9"
+      height="268.5"
+    />
+    <rect x="130.5" y={117} className="adminst9" width="392.1" height="232.6" />
+    <rect x="142.9" y="370.9" className="adminst3" width="63.1" height="11.1" />
+    <rect x="446.5" y="370.7" className="adminst3" width="63.1" height="11.3" />
+    <rect
+      x="128.3"
+      y="116.9"
+      className="adminst10"
+      width="394.2"
+      height="25.3"
+    />
+    <rect
+      x="128.3"
+      y="142.2"
+      className="adminst11"
+      width="394.2"
+      height="210.4"
+    />
+    <rect x="138.7" y="124.9" className="adminst12" width="94.1" height="9.6" />
+    <rect
+      x="128.3"
+      y="142.2"
+      className="adminst12"
+      width="86.5"
+      height="210.4"
+    />
+    <rect x="132.3" y="152.8" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="152.8" className="adminst13" width="57.2" height={7} />
+    <rect x="132.3" y="169.7" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="169.7" className="adminst13" width="50.6" height={7} />
+    <rect x="132.3" y="187.6" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="187.6" className="adminst13" width="53.2" height={7} />
+    <rect x="132.3" y="205.5" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="205.5" className="adminst13" width="50.6" height={7} />
+    <rect x="132.3" y="222.6" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="222.6" className="adminst13" width="38.7" height={7} />
+    <rect x="132.3" y="239.5" className="adminst13" width="12.6" height={7} />
+    <rect x="149.5" y="239.5" className="adminst13" width="53.2" height={7} />
+    <rect
+      x="224.5"
+      y="153.4"
+      className="adminst12"
+      width="288.9"
+      height="165.9"
+    />
+    <path
+      className="adminst14"
+      d="M513.7 319.5H224.2V153.2h289.4V319.5zM224.7 319.1h288.4V153.5H224.7V319.1z"
+    />
+    <rect
+      x="224.7"
+      y="319.7"
+      className="adminst11"
+      width="288.9"
+      height="25.1"
+    />
+    <path
+      className="adminst14"
+      d="M513.9 345H224.5v-25.5h289.4V345zM225 344.6h288.4v-24.7H225V344.6z"
+    />
+    <rect x="235.5" y="161.1" className="adminst15" width="75.5" height="8.6" />
+    <rect x="469.4" y={162} className="adminst10" width="35.4" height="8.6" />
+    <rect x="426.5" y="161.8" className="adminst10" width="35.4" height="8.6" />
+    <rect x={385} y="161.8" className="adminst10" width="35.4" height="8.6" />
+    <rect
+      x="224.5"
+      y="319.2"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="301.7"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="283.7"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="266.2"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="249.6"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="232.1"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect
+      x="224.5"
+      y="216.5"
+      className="adminst16"
+      width="288.9"
+      height="0.2"
+    />
+    <rect x="224.5" y={199} className="adminst16" width="288.9" height="0.2" />
+    <rect x="229.6" y="205.2" className="adminst13" width="61.4" height="6.1" />
+    <rect x="331.3" y="205.2" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="205.2" className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y={222} className="adminst13" width="74.4" height="6.1" />
+    <rect x="331.3" y={222} className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y={222} className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y="237.9" className="adminst13" width="52.8" height="6.1" />
+    <rect x="331.3" y="237.9" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="237.9" className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y="255.1" className="adminst13" width="74.4" height="6.1" />
+    <rect x="331.3" y="255.1" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="255.1" className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y="271.8" className="adminst13" width="61.4" height="6.1" />
+    <rect x="331.3" y="271.8" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="271.8" className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y="289.6" className="adminst13" width="69.7" height="6.1" />
+    <rect x="331.3" y="289.6" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="289.6" className="adminst13" width={10} height="6.1" />
+    <rect x="229.6" y="307.7" className="adminst13" width="74.4" height="6.1" />
+    <rect x="331.3" y="307.7" className="adminst13" width="36.9" height="6.1" />
+    <rect x="426.6" y="307.7" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="205.2" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y={222} className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="237.9" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="255.1" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="271.8" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="289.6" className="adminst13" width={10} height="6.1" />
+    <rect x="403.1" y="307.7" className="adminst13" width={10} height="6.1" />
+    <rect x="490.9" y="205.2" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="205.2" className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y={222} className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y={222} className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y="237.9" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="237.9" className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y="255.1" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="255.1" className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y="271.8" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="271.8" className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y="289.6" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="289.6" className="adminst10" width="13.5" height="6.1" />
+    <rect x="490.9" y="307.7" className="adminst10" width="13.5" height="6.1" />
+    <rect x="467.1" y="307.7" className="adminst10" width="13.5" height="6.1" />
+    <circle className="adminst10" cx="349.7" cy="332.5" r={6} />
+    <circle className="adminst12" cx="365.5" cy="332.5" r={6} />
+    <path
+      className="adminst14"
+      d="M365.5 338.8c-3.5 0-6.3-2.8-6.3-6.3 0-3.5 2.8-6.3 6.3-6.3 3.5 0 6.3 2.8 6.3 6.3C371.8 336 369 338.8 365.5 338.8zM365.5 326.7c-3.2 0-5.8 2.6-5.8 5.8s2.6 5.8 5.8 5.8 5.8-2.6 5.8-5.8S368.7 326.7 365.5 326.7z"
+    />
+    <circle className="adminst12" cx="380.7" cy="332.5" r={6} />
+    <path
+      className="adminst14"
+      d="M380.7 338.8c-3.5 0-6.3-2.8-6.3-6.3 0-3.5 2.8-6.3 6.3-6.3 3.5 0 6.3 2.8 6.3 6.3C387 336 384.2 338.8 380.7 338.8zM380.7 326.7c-3.2 0-5.8 2.6-5.8 5.8s2.6 5.8 5.8 5.8 5.8-2.6 5.8-5.8S383.9 326.7 380.7 326.7z"
+    />
+    <circle className="adminst12" cx="395.6" cy="332.8" r={6} />
+    <path
+      className="adminst14"
+      d="M395.6 339c-3.5 0-6.3-2.8-6.3-6.3 0-3.5 2.8-6.3 6.3-6.3 3.5 0 6.3 2.8 6.3 6.3C401.9 336.2 399.1 339 395.6 339zM395.6 327c-3.2 0-5.8 2.6-5.8 5.8s2.6 5.8 5.8 5.8 5.8-2.6 5.8-5.8S398.8 327 395.6 327z"
+    />
+    <path
+      className="adminst12"
+      d="M363.4 359.2c-0.7 24.6-17.4 43.9-38.4 43.6 -21-0.3-37.4-20.5-36.7-45.1 0.7-24.6 14.1-37.7 35.1-37.4S364.1 334.6 363.4 359.2z"
+    />
+    <path d="M322.3 318c-10.9-0.2-21.3 4.7-29.2 13.5 -7.9 8.8-8.2 14-8.5 26.6 -0.4 12.6 3.5 24.5 10.9 33.5 7.4 9.1 17.5 14.2 28.4 14.4 4.7 0.1 9.1-0.8 13.3-2.4 15.8-6.1 27.2-23.1 27.8-43.9C365.7 333.4 344.2 318.3 322.3 318zM324 400.6c-19.5-0.3-34.8-19.3-34.2-42.4 0.3-11.2 0.1-15.1 7.1-22.9 4-4.5 8.7-7.8 13.8-9.8 3.7-1.4 7.5-2.1 11.4-2.1 9.3 0.1 17.9 4.7 24.1 12.7 6.3 8.1 13.8 12.1 13.4 23.4C359 382.8 343.3 400.8 324 400.6z" />
+    <path
+      className="adminst18"
+      d="M323.3 371.7c4.1 10.6 1.7 20.3-5.4 21.6 -7.1 1.3-16.1-6.2-20.2-16.8 -4.1-10.6-1.7-20.3 5.4-21.6C310.2 353.6 319.2 361.1 323.3 371.7z"
+    />
+    <path
+      className="adminst12"
+      d="M318.4 378.7c1.2 2.9-0.3 5.8-3.4 6.4 -3.1 0.6-6.6-1.2-7.8-4.1 -1.2-2.9 0.3-5.8 3.4-6.4C313.7 374 317.2 375.8 318.4 378.7z"
+    />
+    <path
+      className="adminst12"
+      d="M397.1 364c-0.6 20.7-15.8 37.2-33.9 37 -18.1-0.3-32.3-17.2-31.7-37.9 0.6-20.7 13.1-33.1 31.2-32.9C380.8 330.5 397.7 343.3 397.1 364z"
+    />
+    <path d="M361.7 327.8c-9.5-0.1-18.6 4-25.5 11.6 -6.9 7.5-8.2 13.4-8.5 24.1 -0.3 10.7 3.1 20.9 9.5 28.6 6.5 7.8 15.3 12.1 24.8 12.3 4.1 0.1 8.1-0.7 11.9-2.1 13.9-5.4 24.3-20.2 24.8-37.7C399.3 342.2 381.2 328 361.7 327.8zM393.2 364.4c-0.4 15.1-9.2 27.8-21.1 32.4 -3.1 1.2-6.5 1.8-10 1.8 -8-0.1-15.4-3.8-20.9-10.4 -5.6-6.6-8.5-15.4-8.2-24.6 0.4-15.1 6.6-23.7 18.4-28.2 3.1-1.2 6.5-1.8 10-1.8 8 0.1 17.1 3.3 22.6 9.9C389.6 350 393.5 355.1 393.2 364.4z" />
+    <path
+      className="adminst18"
+      d="M363 375.5c4 8.7 2.1 16.9-4.2 18.4 -6.3 1.4-14.8-4.5-18.8-13.2 -4-8.7-2.1-16.9 4.2-18.4C350.6 360.9 359 366.8 363 375.5z"
+    />
+    <path
+      className="adminst12"
+      d="M357.9 380.7c1.1 2.5-0.2 5-2.9 5.6s-5.7-1-6.8-3.6c-1.1-2.5 0.2-5 2.9-5.6C353.8 376.6 356.8 378.2 357.9 380.7z"
+    />
+    <polygon
+      className="adminst18"
+      points="266.2 510.1 261.1 510.2 224.6 395.1 269.4 412.3 266.1 415.7 230.8 403.9 "
+    />
+    <polygon
+      className="adminst18"
+      points="249.6 546.7 247.5 555 197.6 396.7 273.6 426.7 269.2 430.4 204.6 404.4 "
+    />
+    <polygon
+      className="adminst18"
+      points="305.3 568.7 226.5 447 292.8 452.3 294.8 457.1 237.3 452.8 312.2 568.5 "
+    />
+    <polyline
+      className="adminst18"
+      points="319.1 413.8 321.8 489.2 316.1 488.5 311.9 409 "
+    />
+    <path
+      className="adminst18"
+      d="M315.8 487.8c0 0-6.6-11.3-13.5-3.9 -6.9 7.4-2.6 17.3-2.6 17.3s2.3 3.1 10.1 3.3c7.8 0.2 9.6-0.2 11-2.6 1.4-2.4 1-13.3 1-13.3L315.8 487.8z"
+    />
+    <path
+      className="adminst18"
+      d="M247.6 554c2 3-13.8-15.7-23-7.2 -11.2 10.3-4.3 24.1-4.3 24.1s3.8 3.3 16.4 3.1c16.5-0.3 15.2-4.6 16.2-8.8 1.8-7.1-4.5-21.9-4.5-21.9L247.6 554z"
+    />
+    <path
+      className="adminst18"
+      d="M261.1 507.8c0 0-3.2-10.5-10.6-3.7 -4.4 4.1-4.6 9-4 12.3 0.5 2.6 2.5 4.7 5.1 5.3 1.6 0.3 3.9 0.7 6.9 0.5 10.9-0.7 9.3-4.1 9.3-5.7 0-9.2-3.5-14.8-3.5-14.8L261.1 507.8z"
+    />
+    <path
+      className="adminst18"
+      d="M305.2 567.9c0 0-9.9-16.4-21.3-6.1 -11.4 10.3-2.7 23.2-2.7 23.2s3.9 4 16.8 3.8c16.9-0.3 18.2-8.9 16.9-13.2 -4.5-14.8-7.1-8-7-8.3L305.2 567.9z"
+    />
+    <polygon
+      className="adminst18"
+      points="415 507.4 421 507.3 438 402.1 369 440.3 370.6 449.1 430.4 412.9 "
+    />
+    <polygon
+      className="adminst18"
+      points="413.8 541.9 419 547.3 447.3 469.9 366.9 450.3 361.6 457.7 437.8 474.4 "
+    />
+    <polygon
+      className="adminst18"
+      points="344.6 569.8 355.3 483.5 321.2 458.1 319.4 464.8 348.4 488.8 338.6 569.8 "
+    />
+    <polyline
+      className="adminst18"
+      points="358.5 418.8 358.6 491.3 364.3 493.4 365.6 413.7 "
+    />
+    <path
+      className="adminst18"
+      d="M364.6 492.6c0 0 5.3-11.5 12.4-4.4s3.3 17.2 3.3 17.2 -2.2 3.2-10 3.6 -9.6 0.2-11.1-2.2c-1.5-2.4-0.2-20.2-0.2-20.2L364.6 492.6z"
+    />
+    <path
+      className="adminst18"
+      d="M419.7 546.3c0 0 9.3-18.8 20.9-8.9 6.5 5.6 7.3 12.5 6.8 17.5 -0.5 4.5-3.8 8.2-8.2 9.2 -2.4 0.5-5.6 1-9.7 1 -16.5 0.3-16.2-5.8-17.5-10 -2.1-7.1 2.2-14.2 2.2-14.2L419.7 546.3z"
+    />
+    <path
+      className="adminst18"
+      d="M420.9 505.8c0 0 5.4-13.4 13.1-6.8 4.3 3.7 4.9 8.3 4.5 11.6 -0.3 3-2.5 5.4-5.4 6.1 -1.6 0.4-3.7 0.6-6.4 0.7 -10.9 0.2-11.4-2.5-12.3-5.3 -1.4-4.7 3-16.7 3-16.7L420.9 505.8z"
+    />
+    <path
+      className="adminst18"
+      d="M345.5 564.5c0 0 8.4-15.9 20.2-6 11.8 9.9 5.3 23.9 5.3 23.9s-3.7 4.2-16.7 4.4c-16.9 0.3-17.9-8.1-17.4-12.5 2.4-18.5 5.6-14 5.6-14L345.5 564.5z"
+    />
+    <path
+      className="adminst35"
+      d="M366.6 345.4c41.6 30.2 40 53.4 13 90.6 -21.3 29.3-55.4 38.7-89.1 14.2s-29.3-55.3-8-84.7C303.9 336.2 332.9 320.9 366.6 345.4z"
+    />
+    <path
+      className="adminst18"
+      d="M381.9 438.8c-11.3 15.5-25.7 25.3-41.7 28.4 -17.1 3.3-35.1-1.5-52-13.8 -31.6-22.9-34.7-53.7-9-89.2 26-35.9 57.5-43.2 88.6-20.6 16.8 12.2 31.8 27.3 33.7 45.2C403.3 405.1 393.1 423.4 381.9 438.8zM284.9 368.3c-23.7 32.6-21.4 59.6 7 80.2 30.1 21.9 62.5 16.6 84.5-13.8 28-38.6 25.4-58.9-12-86.1C329 323 301.6 345.2 284.9 368.3z"
+    />
+    <rect
+      x="314.7"
+      y="426.8"
+      transform="matrix(-0.5582 0.8297 -0.8297 -0.5582 854.7977 417.378)"
+      className="adminst56"
+      width="3.2"
+      height={19}
+    />
+    <rect
+      x="314.7"
+      y="426.8"
+      transform="matrix(0.3593 0.9332 -0.9332 0.3593 609.7914 -15.6263)"
+      className="adminst56"
+      width="3.2"
+      height={19}
+    />
+  </svg>
+);
+
+export default Admin;
