+++ client/components/Api/index.js
@@ -1,0 +1,459 @@
+import React from "react";
+
+const Api = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={100}
+    height={100}
+    viewBox="0 0 646 646"
+  >
+    <style jsx>
+      {`
+        .apist0 {
+          fill: #273942;
+        }
+        .apist1 {
+          fill: #b7cdd8;
+        }
+        .apist2 {
+          fill: #e2e2f7;
+        }
+        .apist3 {
+          fill: #52677a;
+        }
+        .apist4 {
+          fill: #354f5c;
+        }
+        .apist5 {
+          fill: #ce521f;
+        }
+        .apist6 {
+          fill: #ebebeb;
+        }
+        .apist7 {
+          fill: #8e96a3;
+        }
+        .apist8 {
+          fill: #3d4d5c;
+        }
+        .apist9 {
+          fill: #3e9697;
+        }
+        .apist10 {
+          fill: #68b7ce;
+        }
+        .apist11 {
+          fill: #e8e8e8;
+        }
+        .apist12 {
+          fill: #ffffff;
+        }
+        .apist13 {
+          fill: #828282;
+        }
+        .apist14 {
+          fill: #cccccc;
+        }
+        .apist15 {
+          fill: #606060;
+        }
+        .apist16 {
+          fill: #c6c6c6;
+        }
+        .apist17 {
+          opacity: 0.15;
+        }
+        .apist18 {
+          fill: #1d1e1c;
+        }
+        .apist19 {
+          fill: #ffd700;
+        }
+        .apist20 {
+          fill: #ff9700;
+        }
+        .apist21 {
+          fill: #ffffe8;
+        }
+        .apist22 {
+          fill: #f2dab8;
+        }
+        .apist23 {
+          fill: #d44d41;
+        }
+        .apist24 {
+          opacity: 0.3;
+        }
+        .apist25 {
+          opacity: 0.3;
+          fill: #ffffff;
+        }
+        .apist26 {
+          fill: #f01c01;
+        }
+        .apist27 {
+          fill: #dbdad9;
+        }
+        .apist28 {
+          opacity: 0.4;
+          fill: #ffffff;
+        }
+        .apist29 {
+          opacity: 0.2;
+        }
+        .apist30 {
+          opacity: 0.2;
+          fill: #020202;
+        }
+        .apist31 {
+          fill: #f3554b;
+        }
+        .apist32 {
+          fill: #bf2c28;
+        }
+        .apist33 {
+          fill: #1f6a7b;
+        }
+        .apist34 {
+          fill: #0594af;
+        }
+        .apist35 {
+          fill: #38a9b4;
+        }
+        .apist36 {
+          opacity: 0.2;
+          fill: #1d1e1c;
+        }
+        .apist37 {
+          opacity: 0.64;
+          fill: #a6d9ed;
+        }
+        .apist38 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #f0f0f0;
+        }
+        .apist39 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #e1dddd;
+        }
+        .apist40 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #ffffff;
+        }
+        .apist41 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #af3737;
+        }
+        .apist42 {
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #990000;
+        }
+        .apist43 {
+          fill: #990000;
+        }
+        .apist44 {
+          fill: #510000;
+        }
+        .apist45 {
+          opacity: 0.5;
+          fill-rule: evenodd;
+          clip-rule: evenodd;
+          fill: #8e2525;
+        }
+        .apist46 {
+          fill: #2ba2b2;
+        }
+        .apist47 {
+          fill: #34aac0;
+        }
+        .apist48 {
+          fill: #f69f00;
+        }
+        .apist49 {
+          fill: #f59e00;
+        }
+        .apist50 {
+          opacity: 0.1;
+        }
+        .apist51 {
+          fill: #ed9406;
+        }
+        .apist52 {
+          fill: #c6006d;
+        }
+        .apist53 {
+          fill: #a50263;
+        }
+        .apist54 {
+          fill: #6ac700;
+        }
+        .apist55 {
+          fill: #5b9904;
+        }
+        .apist56 {
+          fill: #070707;
+        }
+      `}
+    </style>
+    <ellipse className="apist17" cx={349} cy="531.8" rx="65.6" ry="25.2" />
+    <polyline
+      className="apist18"
+      points="258.9 104.3 202.1 24.7 196.6 27.3 257 110.5 "
+    />
+    <path
+      className="apist18"
+      d="M200.5 31.9c0 0-13.3-11.2-4.5-20.9s18.5-2.3 18.5-2.3 2.7 3.8 1.3 15.6c-1.8 15.3-6.7 14.9-10.1 14.6"
+    />
+    <polygon
+      className="apist18"
+      points="318.9 219.3 317.5 223.6 238.2 211.9 281.3 171.1 287.3 174.3 247.7 208.9 "
+    />
+    <path
+      className="apist18"
+      d="M316.4 223.2c0 0-10.9 0.8-7.9 7.8 1.7 3.9 4.9 5.4 7.3 5.9 2.2 0.5 4.4-0.5 5.6-2.5 0.6-1.1 1.3-2.5 2-4.5 2.6-7.7 0.8-8.8-1-10 -3.1-2.1-12.6-1.7-12.6-1.7L316.4 223.2z"
+    />
+    <polygon
+      className="apist18"
+      points="350.4 153.9 353.3 151.6 324.5 95 304.1 138.5 308 142.1 324.6 102.9 "
+    />
+    <path
+      className="apist18"
+      d="M352.8 150.9c0 0-2.1-8.4 4-8 3.4 0.2 5.3 2.2 6.3 3.9 0.9 1.5 0.7 3.5-0.5 4.9 -0.7 0.8-1.6 1.7-2.9 2.7 -5.3 4.1-6.5 3-7.9 1.9 -2.3-1.8-4.5-9.1-4.5-9.1L352.8 150.9z"
+    />
+    <polygon
+      className="apist18"
+      points="337.3 138.6 339 136.3 294.1 91.9 292.1 131.7 296.7 134.2 297 98.6 "
+    />
+    <path
+      className="apist18"
+      d="M338.3 135.8c0 0-4.8-6.9 0.6-7.4 3-0.3 5.4 1.2 6.8 2.5 1.3 1.2 1.8 2.9 1.3 4.2 -0.3 0.7-0.8 1.6-1.5 2.7 -3.1 4.2-4.5 3.4-6.1 2.7 -2.6-1.2-7.1-7.1-7.1-7.1L338.3 135.8z"
+    />
+    <path
+      className="apist19"
+      d="M316.8 507c10.4-23.8-3.1-66.9-3.1-66.9l-46.8-32.7 -46.8-32.7c0 0-45 2.1-63.8 20.1 -18.8 18-41.4 68.1-41.4 68.1l27.2-13.4c0 0-37.1 40.1-36.4 101.5 0.6 61.4-0.5 88-0.5 88s24.6-10.2 82.5-30.7c57.9-20.5 82.8-69.1 82.8-69.1l-3.2 30.1C267 569.3 306.3 530.8 316.8 507z"
+    />
+    <path
+      className="apist20"
+      d="M313.7 440.1c0 0 13.5 43 3.1 66.9S267 569.3 267 569.3l3.2-30.1c0 0-24.9 48.6-82.8 69.1S105 639 105 639l161.9-231.6L313.7 440.1z"
+    />
+    <path
+      className="apist21"
+      d="M266.4 408l-35.1-24.5c0 0-32.5-1-45.1 10.3 -12.6 11.3-26.2 44.3-26.2 44.3l18.9-7.7c0 0-24.6 25.4-20.8 67.4 3.2 35.9 4.2 54.3 4.4 58.9L266.4 408z"
+    />
+    <path
+      className="apist22"
+      d="M266.5 407.9l35.1 24.5c0 0 12.1 30.1 5.8 45.9 -6.2 15.7-32.6 39.9-32.6 39.9l0.7-20.4c0 0-15.3 31.9-56.1 42.7 -34.8 9.3-52.5 14.8-56.8 16.2L266.5 407.9z"
+    />
+    <path
+      className="apist23"
+      d="M439.1 355.3c0 0 16.1 43 14.5 58.4s-15.3 38.7-39.9 63.9c-24.6 25.2-44.1 40-44.1 40l-14.7-81.1L439.1 355.3z"
+    />
+    <path
+      className="apist24"
+      d="M354.9 436.6l3.4 19c13.4-7.3 30.7-20.7 44.2-30.1 13.3-9.3 32.5-36.7 42.8-52.1 -3.3-10.2-6.2-18.1-6.2-18.1L354.9 436.6z"
+    />
+    <path
+      className="apist23"
+      d="M257.3 221.7c0 0-45.9-0.4-59.8 6.5 -13.9 6.9-31.1 27.6-46.3 59.4s-22.4 55.2-22.4 55.2l81.2-14.1L257.3 221.7z"
+    />
+    <path
+      className="apist25"
+      d="M246.8 220.9c-14.3 0.3-40 1.5-49.8 6.4 -13.9 6.9-31.1 27.6-46.3 59.4 -4.3 8.9-7.9 17.2-11 24.5 14.8-18.1 44.4-50.1 53-60.1C200.7 241.9 230.4 228.1 246.8 220.9z"
+    />
+    <path
+      className="apist26"
+      d="M537.5 186.4c12-44.1 4.5-125.7 1.2-155.1 -0.7-5.9-6.3-9.8-12-8.4 -28.8 7-108 28-145.3 54.4 -46 32.6-90.4 79.5-121.7 129 -30.7 48.5-60.8 121.4-60.7 147 0 1 0.2 2 0.5 3 2.3 6.6 13.7 33.2 58.1 64.3 44.4 31 73.3 32.6 80.3 32.5 1 0 2-0.2 3-0.5 24.1-8.7 82.2-62 117.2-107.5C493.9 298.7 522.7 240.8 537.5 186.4z"
+    />
+    <path
+      className="apist23"
+      d="M401.3 66.2c-7.7 3.9-14.7 8-20.7 12.2 -46 32.6-90.4 79.5-121.7 129 -13.4 21.2-26.6 46.9-37.3 71.4 18.3 24.7 45.9 58 77.3 82.7 36 28.3 74.9 42.7 98.7 49.5 21-19.6 43-43.1 59.8-64.8 35.7-46.4 64.6-104.3 79.3-158.6 2-7.2 3.4-15.4 4.4-24.3 -18.2-18-41.6-40-63.8-57.1C451.3 86.1 421.7 73.4 401.3 66.2z"
+    />
+    <path
+      className="apist27"
+      d="M379.4 78.7c-45.3 32.5-88.9 78.8-119.8 127.7 -9.3 14.7-18.5 31.6-26.9 48.8l0 0.1c0 0 9 43.2 76.3 91.8 51.5 37.3 94.9 40.6 112.4 40.2 13.3-13.8 26.1-28.3 36.7-42.2 35.7-46.4 64.6-104.3 79.3-158.6 0.2-0.7 0.4-1.5 0.6-2.3 -18.3-19.7-43.8-44.5-70.9-63.4C432.8 96.7 398.7 84.4 379.4 78.7z"
+    />
+    <path
+      className="apist28"
+      d="M379.3 79.1c-46 32.6-90.4 79.5-121.7 129s-62 124.3-60.6 148.5c0 0 2.2 9 12.5 23 9.1-35.6 29.7-97.3 74-168.3C328.2 139.9 425 73.1 489.2 34.1 453.5 44.3 405.8 60.4 379.3 79.1z"
+    />
+    <path
+      className="apist29"
+      d="M536.7 187.5c-14.8 54.4-43.6 112.3-79.3 158.6s-95.4 100.9-118.6 108c0 0-9.2 1-25.9-3.9 30.3-20.8 81.1-61.3 132.5-127.4 51.8-66.5 81.2-180.3 95.8-254C543.9 105.9 545.2 156.2 536.7 187.5z"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 20.8644 425.6071)"
+      className="apist30"
+      cx="418.8"
+      cy="192.8"
+      rx="58.2"
+      ry="58.2"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 26.0083 422.0111)"
+      className="apist31"
+      cx="417.9"
+      cy="186.1"
+      rx="58.2"
+      ry="58.2"
+    />
+    <path
+      className="apist32"
+      d="M383.6 235.2c-27.1-18.9-33.7-56.4-14.8-83.5 18.9-27.1 56.4-33.7 83.5-14.8s33.7 56.4 14.8 83.5S410.7 254.1 383.6 235.2zM450.3 139.7c-25.6-17.9-60.9-11.6-78.7 14 -17.9 25.6-11.6 60.9 14 78.7 25.6 17.9 60.9 11.6 78.7-14S475.8 157.6 450.3 139.7z"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 25.2004 423.3459)"
+      className="apist33"
+      cx="418.8"
+      cy="187.5"
+      rx="44.9"
+      ry="44.9"
+    />
+    <path
+      className="apist34"
+      d="M454.9 213c3.2-4.6 5.4-9.6 6.7-14.8 -1.5-9.2-6.6-17.8-14.8-23.5 -16.2-11.3-38.6-7.4-49.9 8.8 -9.8 14.1-8.2 32.7 3 44.8C419.2 236.8 442.4 230.9 454.9 213z"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 -85.2533 408.6919)"
+      className="apist30"
+      cx="349.5"
+      cy="286.1"
+      rx="32.1"
+      ry="32.1"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 -82.4164 406.7087)"
+      className="apist31"
+      cx={349}
+      cy="282.4"
+      rx="32.1"
+      ry="32.1"
+    />
+    <path
+      className="apist32"
+      d="M329.6 310.1c-15.3-10.7-19-31.8-8.3-47.1 10.7-15.3 31.8-19 47.1-8.3s19 31.8 8.3 47.1S344.9 320.8 329.6 310.1zM366.4 257.5c-13.7-9.6-32.7-6.2-42.4 7.5 -9.6 13.7-6.2 32.7 7.5 42.4 13.7 9.6 32.7 6.2 42.4-7.5C383.5 286.1 380.2 267.1 366.4 257.5z"
+    />
+    <ellipse
+      transform="matrix(0.5728 -0.8197 0.8197 0.5728 -82.862 407.4448)"
+      className="apist33"
+      cx="349.5"
+      cy="283.2"
+      rx="24.8"
+      ry="24.8"
+    />
+    <path
+      className="apist34"
+      d="M369.4 297.3c1.8-2.5 3-5.3 3.7-8.1 -0.8-5.1-3.6-9.8-8.2-13 -8.9-6.3-21.3-4.1-27.5 4.9 -5.4 7.8-4.5 18 1.7 24.7C349.7 310.4 362.5 307.1 369.4 297.3z"
+    />
+    <path
+      className="apist23"
+      d="M318.9 354.5l6.6-34 -29.7 17.9c0 0-26.3 36.9-35.4 56.8 -9.1 19.9-32.3 64.7-32.3 64.7s34-37.2 49.6-52.6C293.3 391.9 318.9 354.5 318.9 354.5z"
+    />
+    <path
+      className="apist24"
+      d="M325.1 321.1l-6.6 34c0 0-25.6 37.4-41.2 52.8 -15.6 15.4-49.6 52.6-49.6 52.6L325.1 321.1z"
+    />
+    <path
+      className="apist35"
+      d="M292 103.8c24.1 18.7 21.3 48.7 2.1 73.5 -19.2 24.8-46.6 33.8-70.7 15.1s-27.3-51.3-8.1-76.2S267.9 85.1 292 103.8z"
+    />
+    <path
+      className="apist18"
+      d="M295.7 179.4c-10.2 13.1-22.3 21.7-35 24.8 -13.6 3.3-27.2 0.2-39.3-9.2 -22.6-17.5-31.7-49.9-8.5-79.8 23.5-30.3 58.1-30.2 80.4-13 12 9.3 18.4 21.9 18.3 36.5C311.4 151.9 305.8 166.4 295.7 179.4zM216.8 118.3c-21.4 27.6-12.5 56.8 7.7 72.5 21.5 16.7 47.3 11.1 67.1-14.5 20.2-26 19.5-53.4-1.7-69.9C264.7 86.9 232 98.8 216.8 118.3z"
+    />
+    <ellipse
+      transform="matrix(0.4265 -0.9045 0.9045 0.4265 67.3847 293.8434)"
+      className="apist12"
+      cx="265.4"
+      cy="93.8"
+      rx="14.3"
+      ry="26.8"
+    />
+    <path
+      className="apist18"
+      d="M276.9 112.7c-5.7 0-12.1-1.6-18.3-4.5 -6.7-3.2-12.4-7.6-16-12.4 -3.8-5.1-4.9-10.2-3-14.2 1.9-4.1 6.5-6.5 12.9-6.8 6-0.3 13 1.3 19.7 4.5 6.7 3.2 12.4 7.6 16 12.4 3.8 5.1 4.9 10.2 3 14.2s-6.5 6.5-12.9 6.8C277.8 112.7 277.4 112.7 276.9 112.7zM254 78.2c-0.4 0-0.9 0-1.3 0 -5.1 0.3-8.6 2-10 4.8 -1.3 2.9-0.4 6.7 2.6 10.8 3.3 4.4 8.5 8.4 14.7 11.3 6.2 2.9 12.7 4.4 18.1 4.1 5.1-0.3 8.6-2 10-4.8 1.3-2.9 0.4-6.7-2.6-10.8 -3.3-4.4-8.5-8.4-14.7-11.3l0 0C265 79.7 259.1 78.2 254 78.2z"
+    />
+    <ellipse
+      transform="matrix(0.4134 -0.9105 0.9105 0.4134 73.9771 293.7309)"
+      className="apist18"
+      cx={265}
+      cy="89.4"
+      rx="5.5"
+      ry="11.8"
+    />
+    <ellipse
+      transform="matrix(0.5147 -0.8574 0.8574 0.5147 51.5636 268.5486)"
+      className="apist12"
+      cx={263}
+      cy="88.7"
+      rx="2.6"
+      ry="3.5"
+    />
+    <ellipse
+      transform="matrix(0.4265 -0.9045 0.9045 0.4265 40.3973 298.7331)"
+      className="apist12"
+      cx="255.8"
+      cy="117.5"
+      rx="20.2"
+      ry="35.2"
+    />
+    <path
+      className="apist18"
+      d="M269.4 142.9c-7 0-15-1.8-23-5.6 -8.8-4.1-16.1-10-20.7-16.5 -4.8-6.8-6-13.6-3.4-19 2.6-5.5 8.6-8.8 16.8-9.5 7.9-0.6 17.1 1.3 25.9 5.4l0 0c18.7 8.8 29.3 24.5 24.1 35.6C286.2 139.6 278.7 142.9 269.4 142.9zM242.2 95.5c-0.9 0-1.9 0-2.8 0.1 -7 0.6-12 3.2-14 7.6 -2 4.3-0.9 9.9 3.1 15.6 4.2 6.1 11.1 11.5 19.3 15.4 16.7 7.9 33.9 6.8 38.2-2.4 4.4-9.2-5.7-23.2-22.4-31l0 0C256.4 97.4 248.9 95.5 242.2 95.5z"
+    />
+    <ellipse
+      transform="matrix(0.4134 -0.9105 0.9105 0.4134 46.9846 298.4734)"
+      className="apist18"
+      cx="255.2"
+      cy="112.8"
+      rx="7.8"
+      ry={17}
+    />
+    <ellipse
+      transform="matrix(0.5147 -0.8574 0.8574 0.5147 27.4634 268.8198)"
+      className="apist12"
+      cx="251.2"
+      cy="110.2"
+      rx="3.3"
+      ry="4.4"
+    />
+    <polygon
+      className="apist18"
+      points="351.9 255 317.9 159.1 277.6 167.7 282.3 172.8 316 167.9 348.6 260.1 "
+    />
+    <path
+      className="apist18"
+      d="M358.8 264.3c0 0 8.4 10.3-2 17.3s-17.8-2.1-17.8-2.1 -1.8-4.2 1.9-15c3.7-10.8 4.9-13.2 8-14.3C351.9 249 358.8 264.3 358.8 264.3z"
+    />
+    <polygon
+      className="apist18"
+      points="376.3 210.1 341 125.1 306.8 136.4 310.2 140.9 338.4 131.9 373 215.2 "
+    />
+    <path
+      className="apist18"
+      d="M387.5 221.5c0 0 10.4 8.4 1.6 17.4 -8.8 9-17.9 1.7-17.9 1.7s-2.6-3.8-1.2-15.1c1.4-11.3 2.1-13.9 4.8-15.7C377.6 208 387.5 221.5 387.5 221.5z"
+    />
+    <polygon
+      className="apist18"
+      points="308.2 288.9 290.9 223.5 278.9 194 273.5 197.1 287.1 227.8 303.2 291.5 "
+    />
+    <path
+      className="apist18"
+      d="M316 300.4c0 0 8.3 10.4-2.3 17.3 -10.5 6.9-17.8-2.3-17.8-2.3s-1.7-4.3 2.1-15c3.8-10.7 5.1-13.1 8.2-14.2C309.3 285.1 316 300.4 316 300.4z"
+    />
+    <polyline
+      className="apist18"
+      points="260 157.4 146.7 114.1 142.8 119.2 262.4 164.5 "
+    />
+    <path
+      className="apist18"
+      d="M150 121.4c0 0-21.5-2.6-18.9-18.6s17.6-15 17.6-15 5.4 2.1 11.9 15.3c8.6 17.1 3.2 20.1-0.5 22"
+    />
+  </svg>
+);
+
+export default Api;
