+++ client/components/Arrow/index.js
@@ -1,0 +1,28 @@
+import React from "react";
+
+const Arrow = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={26}
+    height={18}
+    viewBox="0 0 25.9 18"
+  >
+    <style jsx>
+      {`
+        .linkst0 {
+          clip-path: url(#SVGID_2_);
+          fill: #38a9b4;
+        }
+      `}
+    </style>
+    <defs>
+      <rect width="25.9" height={18} />
+    </defs>
+    <path
+      className="linkst0"
+      d="M17 0.3c-0.3-0.4-0.9-0.4-1.3 0 -0.3 0.3-0.3 0.9 0 1.3L22.1 8H0.9C0.4 8 0 8.4 0 8.9c0 0.5 0.4 0.9 0.9 0.9h21.2l-6.4 6.4c-0.3 0.4-0.3 0.9 0 1.3 0.4 0.4 0.9 0.4 1.3 0l8-8c0.4-0.3 0.4-0.9 0-1.3L17 0.3zM17 0.3"
+    />
+  </svg>
+);
+
+export default Arrow;
