+++ client/components/Flag/index.js
@@ -1,0 +1,140 @@
+import React from "react";
+
+const Flag = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width={600}
+    height={30}
+    viewBox="0 0 830 42"
+    className="welcome__flag"
+  >
+    <style jsx>
+      {`
+        .flagst0 {
+          fill: #253032;
+        }
+        .flagst1 {
+          fill: #ffffff;
+        }
+        .flagst2 {
+          fill: #c34536;
+        }
+      `}
+    </style>
+    <rect className="flagst0" width={830} height={42} />
+    <path
+      className="flagst1"
+      d="M241.2 27.1v-6.3c0-0.8-0.2-1.3-0.5-1.7 -0.3-0.4-0.8-0.6-1.5-0.6 -0.9 0-1.6 0.3-2 0.8s-0.7 1.3-0.7 2.4v5.4H235v-6.3c0-0.8-0.2-1.3-0.5-1.7 -0.3-0.4-0.8-0.6-1.5-0.6 -0.9 0-1.6 0.3-2 0.8 -0.4 0.5-0.6 1.4-0.6 2.7v5.1h-1.5v-9.6h1.2l0.2 1.3h0.1c0.3-0.5 0.7-0.8 1.2-1.1 0.5-0.3 1.1-0.4 1.7-0.4 1.5 0 2.5 0.5 3 1.6h0.1c0.3-0.5 0.7-0.9 1.2-1.2s1.2-0.4 1.9-0.4c1.1 0 1.9 0.3 2.4 0.8 0.5 0.6 0.8 1.5 0.8 2.7v6.3H241.2z"
+    />
+    <path
+      className="flagst1"
+      d="M251.5 27.1l-0.3-1.4h-0.1c-0.5 0.6-1 1-1.4 1.2 -0.5 0.2-1.1 0.3-1.8 0.3 -1 0-1.7-0.2-2.2-0.7s-0.8-1.2-0.8-2.1c0-1.9 1.6-3 4.7-3.1l1.6-0.1v-0.6c0-0.8-0.2-1.3-0.5-1.7 -0.3-0.4-0.8-0.5-1.6-0.5 -0.8 0-1.7 0.2-2.7 0.7l-0.4-1.1c0.5-0.3 1-0.5 1.6-0.6s1.1-0.2 1.7-0.2c1.1 0 2 0.3 2.6 0.8 0.6 0.5 0.8 1.3 0.8 2.5v6.6H251.5zM248.3 26.1c0.9 0 1.6-0.2 2.1-0.7 0.5-0.5 0.8-1.2 0.8-2.1v-0.9l-1.5 0.1c-1.2 0-2 0.2-2.5 0.5 -0.5 0.3-0.8 0.8-0.8 1.5 0 0.5 0.2 0.9 0.5 1.2S247.7 26.1 248.3 26.1z"
+    />
+    <path
+      className="flagst1"
+      d="M262.2 25.8L262.2 25.8c-0.8 1-1.8 1.5-3.1 1.5 -1.3 0-2.2-0.4-2.9-1.3s-1-2.1-1-3.7 0.4-2.8 1.1-3.7 1.7-1.3 2.9-1.3c1.3 0 2.3 0.5 3 1.4h0.1l-0.1-0.7 0-0.7v-3.9h1.5v13.7h-1.2L262.2 25.8zM259.3 26c1 0 1.7-0.3 2.2-0.8 0.4-0.5 0.7-1.4 0.7-2.6v-0.3c0-1.4-0.2-2.3-0.7-2.9 -0.5-0.6-1.2-0.9-2.2-0.9 -0.9 0-1.5 0.3-2 1 -0.5 0.7-0.7 1.6-0.7 2.8 0 1.2 0.2 2.2 0.7 2.8S258.4 26 259.3 26z"
+    />
+    <path
+      className="flagst1"
+      d="M270.7 27.3c-1.4 0-2.5-0.4-3.4-1.3 -0.8-0.9-1.2-2.1-1.2-3.6 0-1.6 0.4-2.8 1.1-3.7 0.8-0.9 1.8-1.4 3.1-1.4 1.2 0 2.2 0.4 2.9 1.2 0.7 0.8 1.1 1.8 1.1 3.1v0.9h-6.6c0 1.1 0.3 2 0.9 2.6s1.3 0.9 2.3 0.9c1 0 2.1-0.2 3.1-0.7v1.3c-0.5 0.2-1 0.4-1.5 0.5C271.9 27.2 271.4 27.3 270.7 27.3zM270.3 18.5c-0.8 0-1.4 0.3-1.8 0.8 -0.5 0.5-0.7 1.2-0.8 2.1h5c0-0.9-0.2-1.6-0.6-2.1C271.7 18.7 271.1 18.5 270.3 18.5z"
+    />
+    <path
+      className="flagst1"
+      d="M289.3 27.1l-1.8-5.7c-0.1-0.3-0.3-1.1-0.6-2.4h-0.1c-0.2 1-0.4 1.8-0.6 2.4l-1.8 5.6h-1.7l-2.6-9.6h1.5c0.6 2.4 1.1 4.3 1.4 5.5 0.3 1.3 0.5 2.1 0.6 2.6h0.1c0.1-0.3 0.2-0.8 0.3-1.3 0.1-0.5 0.3-1 0.4-1.3l1.8-5.5h1.6l1.7 5.5c0.3 1 0.6 1.9 0.7 2.5h0.1c0-0.2 0.1-0.5 0.2-1s0.7-2.8 1.8-7.1h1.5l-2.7 9.6H289.3z"
+    />
+    <path
+      className="flagst1"
+      d="M295.3 14.8c0-0.3 0.1-0.6 0.2-0.7 0.2-0.2 0.4-0.2 0.6-0.2 0.2 0 0.4 0.1 0.6 0.2s0.3 0.4 0.3 0.7 -0.1 0.6-0.3 0.7c-0.2 0.2-0.4 0.2-0.6 0.2 -0.2 0-0.5-0.1-0.6-0.2C295.4 15.4 295.3 15.2 295.3 14.8zM296.9 27.1h-1.5v-9.6h1.5V27.1z"
+    />
+    <path
+      className="flagst1"
+      d="M303.1 26.1c0.3 0 0.5 0 0.7-0.1 0.2 0 0.4-0.1 0.6-0.1V27c-0.2 0.1-0.4 0.1-0.7 0.2 -0.3 0.1-0.6 0.1-0.8 0.1 -1.9 0-2.8-1-2.8-2.9v-5.7h-1.4v-0.7l1.4-0.6 0.6-2.1h0.8v2.2h2.8v1.1h-2.8v5.7c0 0.6 0.1 1 0.4 1.3S302.6 26.1 303.1 26.1z"
+    />
+    <path
+      className="flagst1"
+      d="M312.9 27.1v-6.2c0-0.8-0.2-1.4-0.5-1.8s-0.9-0.6-1.7-0.6c-1 0-1.8 0.3-2.2 0.8 -0.5 0.6-0.7 1.5-0.7 2.7v5h-1.5V13.4h1.5v4.1c0 0.5 0 0.9-0.1 1.2h0.1c0.3-0.5 0.7-0.8 1.2-1.1 0.5-0.3 1.1-0.4 1.8-0.4 1.2 0 2.1 0.3 2.7 0.8 0.6 0.6 0.9 1.4 0.9 2.7v6.3H312.9z"
+    />
+    <path
+      className="flagst1"
+      d="M364 17.3c1.3 0 2.2 0.4 2.9 1.3 0.7 0.9 1 2.1 1 3.7s-0.4 2.8-1.1 3.7c-0.7 0.9-1.7 1.3-2.9 1.3 -0.6 0-1.2-0.1-1.7-0.3 -0.5-0.2-1-0.6-1.3-1.1h-0.1l-0.3 1.2h-1V13.4h1.5v3.3c0 0.7 0 1.4-0.1 2h0.1C361.6 17.8 362.6 17.3 364 17.3zM363.7 18.5c-1 0-1.7 0.3-2.2 0.9 -0.4 0.6-0.7 1.5-0.7 2.9s0.2 2.3 0.7 2.9c0.5 0.6 1.2 0.9 2.2 0.9 0.9 0 1.6-0.3 2-1 0.4-0.7 0.7-1.6 0.7-2.8 0-1.3-0.2-2.2-0.7-2.8S364.7 18.5 363.7 18.5z"
+    />
+    <path
+      className="flagst1"
+      d="M369 17.5h1.6l2.1 5.5c0.5 1.3 0.8 2.2 0.9 2.7h0.1c0.1-0.3 0.2-0.8 0.5-1.5 0.2-0.7 1-2.9 2.4-6.7h1.6l-4.1 11c-0.4 1.1-0.9 1.9-1.4 2.3 -0.5 0.5-1.2 0.7-2 0.7 -0.4 0-0.9 0-1.3-0.1v-1.2c0.3 0.1 0.7 0.1 1.1 0.1 1 0 1.7-0.6 2.1-1.7l0.5-1.4L369 17.5z"
+    />
+    <path
+      className="flagst2"
+      d="M344.3 11.8c-0.9-0.5-2-0.8-3.1-0.8 -2 0-3.7 0.9-4.9 2.3 -1.2-1.4-2.9-2.3-4.9-2.3 -1.1 0-2.2 0.3-3.1 0.8 -2 1.1-3.3 3.2-3.3 5.6 0 0.7 0.1 1.3 0.3 2 1.1 4.9 11 11.1 11 11.1s9.9-6.2 11-11.1c0.2-0.6 0.3-1.3 0.3-2C347.6 15 346.3 12.9 344.3 11.8L344.3 11.8zM344.3 11.8"
+    />
+    <path
+      className="flagst1"
+      d="M391.1 24.7c0.1-0.1 0.3-0.2 0.5-0.3 0.2-0.2 0.5-0.3 0.9-0.6 0.3-0.2 0.7-0.5 1.1-0.7 0.4-0.3 0.8-0.5 1.1-0.8l0.1-0.8c0.2-1.6 0.4-3.1 0.8-4.5 0.4-1.4 0.8-2.6 1.4-3.7 0.5-1.1 1.1-2.1 1.7-2.9 0.6-0.8 1.2-1.5 1.9-2.1 0.6-0.6 1.3-1 1.9-1.3 0.6-0.3 1.2-0.4 1.7-0.4 0.4 0 0.8 0.1 1.1 0.2 0.3 0.1 0.6 0.4 0.8 0.7 0.2 0.3 0.4 0.7 0.6 1.1s0.2 1 0.2 1.6c0 0.7-0.1 1.5-0.3 2.3 -0.2 0.8-0.5 1.6-0.9 2.4 -0.4 0.8-0.9 1.6-1.4 2.4 -0.6 0.8-1.2 1.6-1.9 2.3 -0.7 0.8-1.4 1.5-2.2 2.1 -0.8 0.7-1.6 1.3-2.5 1.8 0 0.5 0 0.9 0 1.1 0 0.3 0 0.5 0 0.6 0 0.1 0 0.2 0 0.3 0 0.1 0 0.1 0 0.1 0 0.1 0 0.1 0 0.2 0 0.1 0.1 0.1 0.2 0.1 0.1 0 0.2 0.1 0.3 0.1 0.1 0 0.3 0 0.6 0 0.4 0 0.8-0.1 1.4-0.3l3.9-1.1c0.1 0 0.1 0 0.2 0 0.1 0 0.1 0 0.2 0 0.2 0 0.4 0 0.5 0.1s0.3 0.2 0.4 0.3c0.1 0.1 0.2 0.3 0.2 0.4 0 0.2 0.1 0.3 0.1 0.5 0 0.2 0 0.3-0.1 0.5 0 0.1-0.1 0.3-0.2 0.4 -0.1 0.1-0.2 0.2-0.4 0.3 -0.1 0.1-0.3 0.2-0.5 0.2 -0.8 0.3-1.5 0.6-2.3 0.8 -0.3 0.1-0.7 0.2-1.1 0.3 -0.4 0.1-0.8 0.2-1.2 0.3 -0.4 0.1-0.8 0.1-1.2 0.2 -0.4 0-0.8 0.1-1.2 0.1 -0.6 0-1.1-0.1-1.4-0.3 -0.4-0.2-0.7-0.5-0.9-0.8 -0.2-0.3-0.4-0.7-0.5-1.2 -0.1-0.4-0.1-0.9-0.1-1.4 -0.7 0.5-1.2 0.9-1.7 1.2 -0.5 0.3-0.8 0.4-0.9 0.4 -0.1 0-0.3 0-0.4-0.1 -0.1-0.1-0.2-0.2-0.3-0.4 -0.1-0.1-0.2-0.3-0.2-0.5 -0.1-0.2-0.1-0.4-0.1-0.5 0-0.1 0-0.3 0.1-0.4C390.9 24.9 391 24.8 391.1 24.7zM397.8 20.6c0.4-0.4 0.8-0.9 1.2-1.4s0.9-1 1.3-1.5c0.4-0.5 0.8-1.1 1.2-1.6 0.4-0.6 0.8-1.1 1.1-1.7 0.3-0.6 0.6-1.1 0.8-1.6 0.2-0.5 0.4-1 0.4-1.5 0-0.2 0-0.3 0-0.5 0-0.1 0-0.3 0-0.4 0-0.4-0.1-0.8-0.2-1 -0.1-0.2-0.3-0.3-0.5-0.3 -0.2 0-0.5 0.1-0.9 0.4 -0.3 0.3-0.7 0.6-1.1 1.1 -0.4 0.5-0.7 1.1-1.1 1.8s-0.7 1.5-1 2.3c-0.3 0.8-0.6 1.8-0.8 2.8C398 18.5 397.9 19.5 397.8 20.6z"
+    />
+    <path
+      className="flagst1"
+      d="M410.3 28.6c-0.7 0-1.4-0.1-1.9-0.4 -0.6-0.3-1-0.6-1.4-1.1s-0.6-1-0.8-1.6 -0.3-1.2-0.3-1.9c0-0.4 0-0.9 0.1-1.4 0.1-0.5 0.2-1 0.4-1.4 0.2-0.5 0.4-0.9 0.7-1.3 0.3-0.4 0.6-0.8 0.9-1.1 0.4-0.3 0.8-0.6 1.2-0.8 0.4-0.2 0.9-0.3 1.5-0.3 0.4 0 0.8 0.1 1.2 0.2 0.4 0.1 0.7 0.3 1 0.6 0.3 0.3 0.5 0.6 0.7 0.9 0.2 0.4 0.2 0.8 0.2 1.2 0 0.5-0.1 1-0.4 1.5 -0.2 0.5-0.6 0.9-1 1.4 -0.4 0.4-0.9 0.8-1.5 1.1 -0.6 0.3-1.2 0.6-1.9 0.8 0.1 0.2 0.2 0.4 0.3 0.5 0.1 0.1 0.2 0.2 0.3 0.3 0.1 0.1 0.2 0.1 0.4 0.2 0.1 0 0.3 0 0.4 0 0.5 0 1-0.1 1.5-0.3s1-0.4 1.5-0.7c0.5-0.3 1-0.6 1.4-0.9 0.4-0.3 0.8-0.6 1.1-0.9l1.5 1.6c-0.7 0.7-1.4 1.3-2.2 1.9 -0.3 0.2-0.7 0.5-1.1 0.7 -0.4 0.2-0.8 0.4-1.2 0.6 -0.4 0.2-0.9 0.3-1.3 0.4C411.2 28.5 410.7 28.6 410.3 28.6zM408.7 23.1c0.3 0 0.6-0.1 0.9-0.2 0.3-0.1 0.6-0.3 0.9-0.6 0.3-0.2 0.5-0.5 0.7-0.7 0.2-0.3 0.3-0.5 0.3-0.8 0-0.4-0.1-0.6-0.2-0.9 -0.1-0.2-0.3-0.3-0.5-0.3 -0.3 0-0.6 0.1-0.9 0.2 -0.2 0.1-0.4 0.3-0.6 0.5 -0.2 0.2-0.3 0.4-0.4 0.7s-0.2 0.5-0.2 0.8c-0.1 0.3-0.1 0.5-0.1 0.8C408.7 22.7 408.7 22.9 408.7 23.1z"
+    />
+    <path
+      className="flagst1"
+      d="M415.4 24.1c0.4-0.4 0.8-0.7 1.1-0.9 0.3-0.2 0.5-0.5 0.7-0.7 0.2-0.2 0.4-0.4 0.5-0.6 0.2-0.2 0.3-0.5 0.6-0.8 0.2-0.3 0.5-0.6 0.8-1.1 0.3-0.4 0.7-0.9 1.2-1.6 0.1-0.1 0.2-0.3 0.3-0.4 0.1-0.1 0.3-0.3 0.4-0.4 0.2-0.1 0.3-0.2 0.5-0.3 0.2-0.1 0.4-0.1 0.6-0.1 0.4 0 0.7 0 0.9 0.1 0.2 0.1 0.4 0.2 0.4 0.3 0.1 0.1 0.1 0.2 0.2 0.4 0 0.1 0 0.2 0 0.3 0 0.2-0.1 0.4-0.2 0.6 -0.1 0.2-0.3 0.4-0.5 0.7 -0.2 0.2-0.3 0.4-0.5 0.6 -0.2 0.2-0.3 0.3-0.3 0.4 0 0.2 0 0.5 0.1 0.7 0 0.2 0.1 0.4 0.1 0.6 0.1 0.2 0.1 0.4 0.2 0.6 0.1 0.2 0.1 0.5 0.2 0.7 0.1 0.3 0.1 0.6 0.1 1 0 0.4 0.1 0.8 0.1 1.3 0.3-0.2 0.6-0.4 0.9-0.5 0.3-0.2 0.7-0.3 1-0.5 0.4-0.2 0.7-0.3 1.1-0.5 0.4-0.2 0.7-0.4 1.1-0.7l0 2.6c-0.5 0.4-1 0.7-1.7 1.1 -0.6 0.3-1.3 0.6-2 0.9 -0.7 0.2-1.4 0.4-2.1 0.6 -0.7 0.1-1.4 0.2-2 0.2 -0.5 0-0.9 0-1.3-0.1 -0.4-0.1-0.7-0.2-1-0.4 -0.3-0.2-0.5-0.4-0.7-0.7 -0.2-0.3-0.3-0.7-0.3-1.1 -0.1 0-0.2 0-0.3 0 -0.1 0-0.2 0-0.3 0 -0.1 0-0.2 0-0.4 0 -0.1 0-0.2 0-0.3-0.1 -0.1-0.1-0.1-0.1-0.2-0.2 0-0.1-0.1-0.3-0.1-0.5 0-0.2 0-0.4 0.1-0.5 0.1-0.2 0.2-0.3 0.3-0.4s0.2-0.2 0.4-0.3C415.2 24.2 415.3 24.1 415.4 24.1zM416.9 25.6c0.1 0 0.2 0.1 0.3 0.2 0.1 0.1 0.2 0.2 0.4 0.4 0.1 0.1 0.3 0.3 0.5 0.4 0.2 0.1 0.4 0.2 0.7 0.2 0.3 0 0.6-0.1 0.8-0.2 0.2-0.1 0.4-0.2 0.5-0.4 0.1-0.2 0.2-0.4 0.2-0.6 0-0.2 0.1-0.5 0.1-0.7 0-0.2 0-0.4 0-0.6 0-0.2-0.1-0.4-0.1-0.6 0-0.2-0.1-0.4-0.1-0.6 -0.1-0.2-0.1-0.3-0.1-0.5L416.9 25.6z"
+    />
+    <path
+      className="flagst1"
+      d="M433.6 13c0-0.7 0.1-1.2 0.4-1.7 0.3-0.5 0.6-0.9 1.1-1.2 0.5-0.3 1.1-0.6 1.8-0.8 0.7-0.2 1.4-0.4 2.3-0.5 0.8-0.1 1.7-0.2 2.6-0.3s1.9-0.1 2.9-0.1c1.4 0 2.8 0.1 4.3 0.2 1.5 0.1 3 0.3 4.7 0.6 0.4 0.1 0.6 0.2 0.8 0.4 0.2 0.2 0.2 0.4 0.2 0.7 0 0.2 0 0.3-0.1 0.5 -0.1 0.1-0.2 0.3-0.3 0.4 -0.1 0.1-0.3 0.2-0.4 0.3 -0.2 0.1-0.3 0.1-0.5 0.1 -0.6 0-1.2 0-1.9-0.1 -0.7-0.1-1.3-0.1-2-0.2 -0.7-0.1-1.4-0.2-2.2-0.2 -0.8-0.1-1.5-0.1-2.3-0.2 -0.1 0.9-0.3 1.8-0.5 2.8 -0.2 1-0.5 2-0.7 3.1s-0.5 2.1-0.8 3.1c-0.3 1-0.6 2-0.8 3 -0.3 0.9-0.5 1.8-0.7 2.6 -0.2 0.8-0.4 1.5-0.6 2.1 -0.1 0.3-0.2 0.5-0.3 0.7 -0.1 0.2-0.3 0.4-0.4 0.6 -0.2 0.2-0.3 0.3-0.5 0.4 -0.2 0.1-0.3 0.1-0.5 0.1 -0.4 0-0.8-0.1-1.1-0.3 -0.3-0.2-0.4-0.6-0.4-1 0-0.1 0-0.3 0.1-0.5 0.1-0.7 0.2-1.4 0.4-2.1 0.2-0.7 0.4-1.4 0.6-2.2 0.2-0.7 0.5-1.5 0.8-2.4 0.3-0.8 0.6-1.8 1-2.8s0.7-2.1 1.1-3.3c0.4-1.2 0.9-2.6 1.3-4 -0.4 0-0.8 0.1-1.3 0.1 -0.4 0-0.8 0.1-1.2 0.2 -0.4 0.1-0.8 0.2-1.1 0.3 -0.3 0.1-0.7 0.2-0.9 0.4 -0.3 0.2-0.5 0.3-0.6 0.6 -0.2 0.2-0.2 0.5-0.2 0.8 0 0 0 0.1 0 0.3 0 0.2-0.1 0.3-0.2 0.5 -0.1 0.2-0.3 0.4-0.6 0.5 -0.3 0.2-0.6 0.2-1.1 0.2 -0.5 0-0.8-0.1-1.1-0.2 -0.3-0.2-0.4-0.3-0.6-0.5C433.7 13.5 433.6 13.3 433.6 13z"
+    />
+    <path
+      className="flagst1"
+      d="M458.1 26.1c-0.3 0.3-0.6 0.5-1 0.8 -0.4 0.3-0.9 0.6-1.4 0.8 -0.5 0.2-1 0.5-1.5 0.6 -0.5 0.2-1 0.2-1.4 0.2 -0.4 0-0.7-0.1-1-0.2 -0.3-0.1-0.5-0.3-0.7-0.5 -0.2-0.2-0.3-0.5-0.4-0.7 -0.1-0.3-0.1-0.6-0.1-1 0-0.5 0.1-1.1 0.2-1.7 0.1-0.7 0.3-1.3 0.4-2 0.2-0.7 0.3-1.4 0.5-2.1 0.2-0.7 0.3-1.3 0.5-1.9 0.1-0.2 0.2-0.4 0.3-0.5 0.2-0.1 0.3-0.3 0.6-0.4 0.2-0.1 0.4-0.2 0.6-0.2 0.2 0 0.4-0.1 0.6-0.1 0.3 0 0.6 0.1 0.7 0.2 0.1 0.1 0.2 0.3 0.2 0.5 0 0.2 0 0.4-0.1 0.7 -0.1 0.3-0.2 0.7-0.3 1.1 -0.1 0.4-0.2 0.9-0.4 1.3 -0.1 0.5-0.3 0.9-0.4 1.4 -0.1 0.5-0.2 0.9-0.3 1.4s-0.1 0.9-0.1 1.2c0 0.4 0 0.6 0.1 0.8 0.1 0.2 0.2 0.3 0.5 0.3 0.3 0 0.6-0.1 0.9-0.3 0.3-0.2 0.7-0.4 1-0.7 0.4-0.3 0.7-0.6 1-0.9 0.3-0.3 0.6-0.6 0.9-0.8V26.1zM452.9 14c0-0.2 0-0.4 0.1-0.6 0.1-0.2 0.2-0.4 0.4-0.6 0.2-0.2 0.4-0.4 0.7-0.5 0.3-0.1 0.7-0.2 1.1-0.2 0.3 0 0.5 0 0.8 0.1 0.2 0.1 0.5 0.2 0.6 0.3 0.2 0.1 0.3 0.3 0.4 0.5 0.1 0.2 0.2 0.5 0.2 0.8 0 0.3 0 0.5-0.1 0.8 -0.1 0.2-0.2 0.5-0.4 0.7s-0.4 0.4-0.7 0.5c-0.3 0.1-0.6 0.2-1 0.2 -0.7 0-1.2-0.2-1.5-0.5C453.1 15 452.9 14.6 452.9 14z"
+    />
+    <path
+      className="flagst1"
+      d="M460.2 28.6c-0.4 0-0.8-0.2-1.2-0.6 -0.4-0.4-0.7-0.9-0.9-1.5 -0.3-0.6-0.5-1.3-0.6-2.2 -0.2-0.8-0.2-1.7-0.2-2.6 0-0.7 0.1-1.5 0.2-2.3 0.1-0.8 0.2-1.7 0.4-2.5 0.2-0.9 0.4-1.7 0.7-2.6 0.3-0.9 0.6-1.7 0.9-2.5 0.3-0.8 0.7-1.6 1.1-2.3 0.4-0.7 0.8-1.3 1.3-1.8 0.4-0.5 0.9-0.9 1.4-1.2 0.5-0.3 1-0.4 1.5-0.4 0.3 0 0.6 0.1 0.8 0.2 0.2 0.2 0.4 0.4 0.5 0.6s0.2 0.5 0.3 0.8c0.1 0.3 0.1 0.6 0.1 1 0 0.7-0.1 1.5-0.3 2.2 -0.2 0.8-0.4 1.5-0.7 2.3 -0.3 0.8-0.6 1.5-1 2.3s-0.8 1.5-1.2 2.3c-0.4 0.8-0.8 1.5-1.2 2.2 -0.4 0.7-0.8 1.5-1.2 2.2 0 0.1-0.1 0.2-0.1 0.3 0 0.1-0.1 0.3-0.1 0.4 0 0.1-0.1 0.3-0.1 0.5 0 0.2-0.1 0.3-0.1 0.5v0.2c0 0.2 0 0.4 0.1 0.6 0 0.2 0.1 0.4 0.2 0.5 0.1 0.2 0.1 0.3 0.2 0.4 0.1 0.1 0.1 0.1 0.2 0.1 0.1 0 0.3 0 0.5-0.1s0.4-0.1 0.6-0.3c0.2-0.1 0.4-0.2 0.7-0.4 0.2-0.1 0.4-0.3 0.7-0.4 0.5-0.3 1-0.7 1.5-1.1l1 2.5c-0.7 0.5-1.4 0.9-2.1 1.3 -0.3 0.2-0.6 0.3-0.9 0.5 -0.3 0.2-0.6 0.3-1 0.4 -0.3 0.1-0.6 0.2-0.9 0.3C460.7 28.5 460.4 28.6 460.2 28.6zM460.2 19.2c0.4-0.6 0.7-1.3 1.1-2.1 0.4-0.8 0.8-1.6 1.1-2.4 0.4-0.8 0.7-1.7 0.9-2.5 0.3-0.8 0.4-1.6 0.5-2.3 0-0.2 0-0.4 0-0.6 0-0.2 0-0.2-0.1-0.2 -0.2 0-0.5 0.1-0.7 0.4 -0.3 0.3-0.5 0.7-0.8 1.2s-0.5 1.1-0.8 1.7 -0.5 1.4-0.6 2.1 -0.3 1.5-0.4 2.3C460.2 17.7 460.2 18.4 460.2 19.2z"
+    />
+    <path
+      className="flagst1"
+      d="M467.1 28.6c-0.4 0-0.8-0.2-1.2-0.6 -0.4-0.4-0.7-0.9-0.9-1.5 -0.3-0.6-0.5-1.3-0.6-2.2 -0.2-0.8-0.2-1.7-0.2-2.6 0-0.7 0.1-1.5 0.2-2.3 0.1-0.8 0.2-1.7 0.4-2.5 0.2-0.9 0.4-1.7 0.7-2.6 0.3-0.9 0.6-1.7 0.9-2.5 0.3-0.8 0.7-1.6 1.1-2.3 0.4-0.7 0.8-1.3 1.3-1.8 0.4-0.5 0.9-0.9 1.4-1.2 0.5-0.3 1-0.4 1.5-0.4 0.3 0 0.6 0.1 0.8 0.2 0.2 0.2 0.4 0.4 0.5 0.6s0.2 0.5 0.3 0.8 0.1 0.6 0.1 1c0 0.7-0.1 1.5-0.3 2.2 -0.2 0.8-0.4 1.5-0.7 2.3 -0.3 0.8-0.6 1.5-1 2.3s-0.8 1.5-1.2 2.3c-0.4 0.8-0.8 1.5-1.2 2.2 -0.4 0.7-0.8 1.5-1.2 2.2 0 0.1-0.1 0.2-0.1 0.3 0 0.1-0.1 0.3-0.1 0.4 0 0.1-0.1 0.3-0.1 0.5 0 0.2-0.1 0.3-0.1 0.5v0.2c0 0.2 0 0.4 0.1 0.6 0 0.2 0.1 0.4 0.2 0.5 0.1 0.2 0.1 0.3 0.2 0.4 0.1 0.1 0.1 0.1 0.2 0.1 0.1 0 0.3 0 0.5-0.1 0.2-0.1 0.4-0.1 0.6-0.3 0.2-0.1 0.4-0.2 0.7-0.4 0.2-0.1 0.4-0.3 0.7-0.4 0.5-0.3 1-0.7 1.5-1.1l1 2.5c-0.7 0.5-1.4 0.9-2.1 1.3 -0.3 0.2-0.6 0.3-0.9 0.5 -0.3 0.2-0.6 0.3-1 0.4 -0.3 0.1-0.6 0.2-0.9 0.3C467.7 28.5 467.4 28.6 467.1 28.6zM467.2 19.2c0.4-0.6 0.7-1.3 1.1-2.1 0.4-0.8 0.8-1.6 1.1-2.4 0.4-0.8 0.7-1.7 0.9-2.5 0.3-0.8 0.4-1.6 0.5-2.3 0-0.2 0-0.4 0-0.6 0-0.2 0-0.2-0.1-0.2 -0.2 0-0.5 0.1-0.7 0.4 -0.3 0.3-0.5 0.7-0.8 1.2 -0.3 0.5-0.5 1.1-0.8 1.7 -0.2 0.7-0.5 1.4-0.6 2.1s-0.3 1.5-0.4 2.3C467.2 17.7 467.2 18.4 467.2 19.2z"
+    />
+    <path
+      className="flagst1"
+      d="M475.5 28.6c-0.7 0-1.4-0.1-1.9-0.4 -0.6-0.3-1-0.6-1.4-1.1 -0.4-0.4-0.6-1-0.8-1.6 -0.2-0.6-0.3-1.2-0.3-1.9 0-0.4 0-0.9 0.1-1.4 0.1-0.5 0.2-1 0.4-1.4 0.2-0.5 0.4-0.9 0.7-1.3 0.3-0.4 0.6-0.8 0.9-1.1 0.4-0.3 0.8-0.6 1.2-0.8 0.4-0.2 0.9-0.3 1.5-0.3 0.4 0 0.8 0.1 1.2 0.2 0.4 0.1 0.7 0.3 1 0.6 0.3 0.3 0.5 0.6 0.7 0.9 0.2 0.4 0.2 0.8 0.2 1.2 0 0.5-0.1 1-0.4 1.5 -0.2 0.5-0.6 0.9-1 1.4 -0.4 0.4-0.9 0.8-1.5 1.1 -0.6 0.3-1.2 0.6-1.9 0.8 0.1 0.2 0.2 0.4 0.3 0.5 0.1 0.1 0.2 0.2 0.3 0.3 0.1 0.1 0.2 0.1 0.4 0.2 0.1 0 0.3 0 0.4 0 0.5 0 1-0.1 1.5-0.3 0.5-0.2 1-0.4 1.5-0.7 0.5-0.3 1-0.6 1.4-0.9 0.4-0.3 0.8-0.6 1.1-0.9l1.5 1.6c-0.7 0.7-1.4 1.3-2.2 1.9 -0.3 0.2-0.7 0.5-1.1 0.7 -0.4 0.2-0.8 0.4-1.2 0.6 -0.4 0.2-0.9 0.3-1.3 0.4C476.4 28.5 475.9 28.6 475.5 28.6zM473.9 23.1c0.3 0 0.6-0.1 0.9-0.2 0.3-0.1 0.6-0.3 0.9-0.6 0.3-0.2 0.5-0.5 0.7-0.7 0.2-0.3 0.3-0.5 0.3-0.8 0-0.4-0.1-0.6-0.2-0.9 -0.1-0.2-0.3-0.3-0.5-0.3 -0.3 0-0.6 0.1-0.9 0.2s-0.4 0.3-0.6 0.5c-0.2 0.2-0.3 0.4-0.4 0.7 -0.1 0.3-0.2 0.5-0.2 0.8 -0.1 0.3-0.1 0.5-0.1 0.8C473.9 22.7 473.9 22.9 473.9 23.1z"
+    />
+    <path
+      className="flagst1"
+      d="M489.7 18.1c0 0.2 0 0.5-0.1 0.9 0 0.4-0.1 0.8-0.2 1.2 -0.1 0.4-0.2 0.9-0.2 1.3 -0.1 0.5-0.2 0.9-0.2 1.3 -0.1 0.4-0.1 0.7-0.2 1 0 0.3-0.1 0.5-0.1 0.6 0 0.2 0 0.4 0 0.6 0 0.2 0 0.3 0.1 0.5 0 0.1 0.1 0.2 0.2 0.3s0.2 0.1 0.4 0.1c0.2 0 0.4 0 0.6-0.1 0.2-0.1 0.5-0.2 0.8-0.3 0.3-0.1 0.6-0.3 0.8-0.5 0.3-0.2 0.6-0.4 0.8-0.6 0.3-0.2 0.5-0.4 0.7-0.6 0.2-0.2 0.4-0.4 0.5-0.6 0 0.3 0 0.5 0 0.7 0 0.2 0 0.4 0 0.6 0 0.2 0 0.4 0 0.6 0 0.2 0 0.4 0 0.7 -0.2 0.2-0.4 0.4-0.6 0.6 -0.3 0.2-0.5 0.4-0.8 0.6 -0.3 0.2-0.6 0.4-1 0.6 -0.3 0.2-0.7 0.3-1 0.5 -0.4 0.1-0.7 0.2-1 0.3 -0.3 0.1-0.7 0.1-1 0.1 -0.3 0-0.6-0.1-0.8-0.2 -0.2-0.1-0.4-0.3-0.5-0.5 -0.2-0.2-0.3-0.5-0.4-0.8 -0.1-0.3-0.2-0.6-0.3-1 -0.3 0.4-0.6 0.7-0.9 1s-0.6 0.6-1 0.8c-0.3 0.2-0.7 0.4-1 0.5 -0.3 0.1-0.7 0.2-1 0.2 -0.3 0-0.5 0-0.7-0.1 -0.2-0.1-0.4-0.2-0.6-0.4 -0.2-0.2-0.3-0.5-0.4-0.8 -0.1-0.3-0.2-0.7-0.2-1.2 0-0.6 0.1-1.4 0.2-2.3 0.1-0.9 0.3-2 0.5-3.2 0-0.1 0-0.3 0.1-0.5 0-0.2 0.1-0.4 0.2-0.7 0.1-0.2 0.2-0.5 0.3-0.7 0.1-0.2 0.2-0.5 0.4-0.7 0.1-0.2 0.3-0.4 0.5-0.5 0.2-0.1 0.4-0.2 0.7-0.2 0.4 0 0.7 0.1 0.9 0.2 0.2 0.1 0.3 0.2 0.3 0.3 -0.1 0.4-0.1 0.9-0.2 1.4 -0.1 0.5-0.2 1-0.3 1.5 -0.1 0.5-0.2 1-0.2 1.5 -0.1 0.5-0.1 0.9-0.2 1.3 -0.1 0.4-0.1 0.7-0.1 0.9 0 0.2-0.1 0.4-0.1 0.4 0 0.1 0 0.3 0 0.4 0 0.1 0.1 0.2 0.1 0.3 0.1 0.1 0.1 0.2 0.2 0.2 0.1 0.1 0.2 0.1 0.3 0.1 0.2 0 0.3 0 0.5-0.1 0.2-0.1 0.3-0.2 0.5-0.4 0.2-0.2 0.3-0.4 0.4-0.6 0.1-0.2 0.2-0.5 0.3-0.8l1.2-6c0-0.2 0.2-0.3 0.5-0.5 0.3-0.1 0.6-0.2 0.9-0.2 0.4 0 0.7 0.1 0.9 0.2C489.6 17.7 489.7 17.9 489.7 18.1z"
+    />
+    <path
+      className="flagst1"
+      d="M495.1 28.6c-0.4 0-0.8-0.2-1.2-0.6 -0.4-0.4-0.7-0.9-0.9-1.5 -0.3-0.6-0.5-1.3-0.6-2.2 -0.2-0.8-0.2-1.7-0.2-2.6 0-0.7 0.1-1.5 0.2-2.3 0.1-0.8 0.2-1.7 0.4-2.5 0.2-0.9 0.4-1.7 0.7-2.6 0.3-0.9 0.6-1.7 0.9-2.5 0.3-0.8 0.7-1.6 1.1-2.3 0.4-0.7 0.8-1.3 1.3-1.8 0.4-0.5 0.9-0.9 1.4-1.2 0.5-0.3 1-0.4 1.5-0.4 0.3 0 0.6 0.1 0.8 0.2 0.2 0.2 0.4 0.4 0.5 0.6s0.2 0.5 0.3 0.8 0.1 0.6 0.1 1c0 0.7-0.1 1.5-0.3 2.2 -0.2 0.8-0.4 1.5-0.7 2.3 -0.3 0.8-0.6 1.5-1 2.3s-0.8 1.5-1.2 2.3c-0.4 0.8-0.8 1.5-1.2 2.2 -0.4 0.7-0.8 1.5-1.2 2.2 0 0.1-0.1 0.2-0.1 0.3 0 0.1-0.1 0.3-0.1 0.4 0 0.1-0.1 0.3-0.1 0.5 0 0.2-0.1 0.3-0.1 0.5v0.2c0 0.2 0 0.4 0.1 0.6 0 0.2 0.1 0.4 0.2 0.5 0.1 0.2 0.1 0.3 0.2 0.4 0.1 0.1 0.1 0.1 0.2 0.1 0.1 0 0.3 0 0.5-0.1 0.2-0.1 0.4-0.1 0.6-0.3 0.2-0.1 0.4-0.2 0.7-0.4 0.2-0.1 0.4-0.3 0.7-0.4 0.5-0.3 1-0.7 1.5-1.1l1 2.5c-0.7 0.5-1.4 0.9-2.1 1.3 -0.3 0.2-0.6 0.3-0.9 0.5 -0.3 0.2-0.6 0.3-1 0.4 -0.3 0.1-0.6 0.2-0.9 0.3C495.7 28.5 495.4 28.6 495.1 28.6zM495.1 19.2c0.4-0.6 0.7-1.3 1.1-2.1 0.4-0.8 0.8-1.6 1.1-2.4 0.4-0.8 0.7-1.7 0.9-2.5 0.3-0.8 0.4-1.6 0.5-2.3 0-0.2 0-0.4 0-0.6 0-0.2 0-0.2-0.1-0.2 -0.2 0-0.5 0.1-0.7 0.4 -0.3 0.3-0.5 0.7-0.8 1.2 -0.3 0.5-0.5 1.1-0.8 1.7 -0.2 0.7-0.5 1.4-0.6 2.1s-0.3 1.5-0.4 2.3C495.2 17.7 495.1 18.4 495.1 19.2z"
+    />
+    <path
+      className="flagst1"
+      d="M499.1 24.1c0.4-0.4 0.8-0.7 1.1-0.9 0.3-0.2 0.5-0.5 0.7-0.7 0.2-0.2 0.4-0.4 0.5-0.6 0.2-0.2 0.3-0.5 0.6-0.8 0.2-0.3 0.5-0.6 0.8-1.1 0.3-0.4 0.7-0.9 1.2-1.6 0.1-0.1 0.2-0.3 0.3-0.4 0.1-0.1 0.3-0.3 0.4-0.4 0.2-0.1 0.3-0.2 0.5-0.3 0.2-0.1 0.4-0.1 0.6-0.1 0.4 0 0.7 0 0.9 0.1 0.2 0.1 0.4 0.2 0.4 0.3 0.1 0.1 0.1 0.2 0.2 0.4 0 0.1 0 0.2 0 0.3 0 0.2-0.1 0.4-0.2 0.6 -0.1 0.2-0.3 0.4-0.5 0.7 -0.2 0.2-0.3 0.4-0.5 0.6 -0.2 0.2-0.3 0.3-0.3 0.4 0 0.2 0 0.5 0.1 0.7 0 0.2 0.1 0.4 0.1 0.6 0.1 0.2 0.1 0.4 0.2 0.6 0.1 0.2 0.1 0.5 0.2 0.7 0.1 0.3 0.1 0.6 0.1 1 0 0.4 0.1 0.8 0.1 1.3 0.3-0.2 0.6-0.4 0.9-0.5 0.3-0.2 0.7-0.3 1-0.5 0.4-0.2 0.7-0.3 1.1-0.5 0.4-0.2 0.7-0.4 1.1-0.7l0 2.6c-0.5 0.4-1 0.7-1.7 1.1 -0.6 0.3-1.3 0.6-2 0.9s-1.4 0.4-2.1 0.6c-0.7 0.1-1.4 0.2-2 0.2 -0.5 0-0.9 0-1.3-0.1 -0.4-0.1-0.7-0.2-1-0.4 -0.3-0.2-0.5-0.4-0.7-0.7 -0.2-0.3-0.3-0.7-0.3-1.1 -0.1 0-0.2 0-0.3 0 -0.1 0-0.2 0-0.3 0 -0.1 0-0.2 0-0.4 0 -0.1 0-0.2 0-0.3-0.1s-0.1-0.1-0.2-0.2c0-0.1-0.1-0.3-0.1-0.5 0-0.2 0-0.4 0.1-0.5 0.1-0.2 0.2-0.3 0.3-0.4s0.2-0.2 0.4-0.3C498.9 24.2 499 24.1 499.1 24.1zM500.6 25.6c0.1 0 0.2 0.1 0.3 0.2 0.1 0.1 0.2 0.2 0.4 0.4s0.3 0.3 0.5 0.4c0.2 0.1 0.4 0.2 0.7 0.2 0.3 0 0.6-0.1 0.8-0.2 0.2-0.1 0.4-0.2 0.5-0.4 0.1-0.2 0.2-0.4 0.2-0.6 0-0.2 0.1-0.5 0.1-0.7 0-0.2 0-0.4 0-0.6 0-0.2-0.1-0.4-0.1-0.6s-0.1-0.4-0.1-0.6c-0.1-0.2-0.1-0.3-0.1-0.5L500.6 25.6z"
+    />
+    <path
+      className="flagst1"
+      d="M516.2 26.3c0 0 0 0.1 0 0.2 0 0.1 0 0.3-0.1 0.4 0 0.2-0.1 0.3-0.2 0.5 -0.1 0.2-0.2 0.4-0.4 0.5 -0.2 0.2-0.4 0.3-0.6 0.4 -0.2 0.1-0.6 0.2-0.9 0.2 -0.3 0-0.6 0-0.8-0.1 -0.2-0.1-0.4-0.2-0.6-0.3 -0.2-0.2-0.3-0.3-0.4-0.6 -0.1-0.2-0.1-0.5-0.1-0.9 0-0.3 0-0.6 0.1-0.8 0.1-0.3 0.2-0.5 0.4-0.7 0.2-0.2 0.4-0.4 0.7-0.5 0.3-0.1 0.6-0.2 1-0.2 0.6 0 1.1 0.2 1.5 0.5C516 25.2 516.2 25.7 516.2 26.3z"
+    />
+    <path
+      className="flagst1"
+      d="M522 28.6c-0.7 0-1.3-0.1-1.9-0.4 -0.5-0.3-1-0.6-1.4-1.1 -0.4-0.4-0.6-1-0.8-1.5 -0.2-0.6-0.3-1.2-0.3-1.8 0-0.5 0-0.9 0.1-1.4 0.1-0.5 0.2-1 0.4-1.4 0.2-0.5 0.4-0.9 0.7-1.4 0.3-0.4 0.6-0.8 0.9-1.1 0.4-0.3 0.8-0.6 1.2-0.8 0.4-0.2 0.9-0.3 1.4-0.3 0.4 0 0.9 0.1 1.2 0.2 0.4 0.1 0.7 0.3 1 0.6 0.3 0.2 0.5 0.5 0.6 0.9 0.2 0.3 0.2 0.7 0.2 1.1 0 0.3 0 0.5-0.1 0.8 -0.1 0.3-0.2 0.5-0.3 0.7 -0.1 0.2-0.3 0.4-0.5 0.5 -0.2 0.1-0.4 0.2-0.7 0.2 -0.4 0-0.6-0.1-0.7-0.3 -0.1-0.2-0.2-0.5-0.2-0.8 0-0.1 0-0.3 0-0.4 0-0.2 0-0.3 0-0.4 0-0.2 0-0.3-0.1-0.5 0-0.1-0.1-0.2-0.3-0.2 -0.3 0-0.6 0.1-0.8 0.2 -0.2 0.2-0.4 0.4-0.6 0.6 -0.2 0.2-0.3 0.5-0.4 0.8 -0.1 0.3-0.2 0.6-0.2 0.9 -0.1 0.3-0.1 0.6-0.1 0.9 0 0.3 0 0.5 0 0.7 0 0.3 0 0.6 0.1 0.9 0.1 0.3 0.1 0.6 0.3 0.8 0.1 0.3 0.3 0.5 0.6 0.6 0.2 0.2 0.6 0.3 0.9 0.3 0.4 0 0.8-0.1 1.2-0.3 0.5-0.2 0.9-0.5 1.4-0.9 0.5-0.4 0.9-0.7 1.3-1.2 0.4-0.4 0.8-0.8 1.1-1.2l0.8 1.4c-0.6 0.8-1.2 1.6-1.9 2.2 -0.3 0.3-0.6 0.6-0.9 0.8 -0.3 0.3-0.7 0.5-1.1 0.7 -0.4 0.2-0.8 0.4-1.2 0.5C522.8 28.5 522.4 28.6 522 28.6z"
+    />
+    <path
+      className="flagst1"
+      d="M534.4 19.8c0.2 0.5 0.4 0.9 0.5 1.4 0.1 0.5 0.2 1 0.2 1.4 0 0.5-0.1 0.9-0.2 1.4 -0.1 0.5-0.2 0.9-0.4 1.4s-0.4 0.8-0.7 1.2c-0.3 0.4-0.6 0.7-0.9 1 -0.4 0.3-0.7 0.5-1.2 0.7 -0.4 0.2-0.9 0.2-1.3 0.2 -0.7 0-1.3-0.2-1.8-0.5 -0.5-0.3-0.9-0.7-1.2-1.2 -0.3-0.5-0.5-1-0.7-1.6 -0.1-0.6-0.2-1.2-0.2-1.9 0-0.9 0.1-1.7 0.4-2.4 0.3-0.7 0.6-1.4 1.1-1.9 0.5-0.6 1.1-1 1.8-1.3 0.7-0.3 1.5-0.5 2.3-0.5 0.7 0 1.4 0.1 2.1 0.3 0.6 0.2 1.3 0.5 1.8 0.9 0.6 0.4 1.1 0.8 1.6 1.3 0.5 0.5 0.9 1 1.4 1.5l-0.2 1.7c-0.2-0.3-0.5-0.6-0.9-0.9 -0.4-0.3-0.7-0.6-1.1-1 -0.4-0.3-0.8-0.6-1.2-0.9 -0.4-0.3-0.7-0.5-1-0.7L534.4 19.8zM530.8 26.3c0.3 0 0.5-0.1 0.8-0.4 0.2-0.3 0.5-0.6 0.7-1 0.2-0.4 0.3-0.9 0.5-1.4 0.1-0.5 0.2-1 0.2-1.4 0-0.4 0-0.7-0.1-1s-0.1-0.6-0.2-0.8c-0.1-0.2-0.2-0.4-0.4-0.5 -0.2-0.1-0.4-0.2-0.6-0.2 -0.3 0-0.6 0.1-0.9 0.4 -0.3 0.3-0.5 0.6-0.8 1 -0.2 0.4-0.4 0.9-0.5 1.4 -0.1 0.5-0.2 1-0.2 1.5 0 0.3 0 0.6 0.1 0.9 0.1 0.3 0.2 0.5 0.3 0.8 0.1 0.2 0.3 0.4 0.5 0.5C530.2 26.2 530.5 26.3 530.8 26.3z"
+    />
+    <path
+      className="flagst1"
+      d="M544.9 19.8c0.2 0.5 0.4 0.9 0.5 1.4 0.1 0.5 0.2 1 0.2 1.4 0 0.5-0.1 0.9-0.2 1.4 -0.1 0.5-0.2 0.9-0.4 1.4s-0.4 0.8-0.7 1.2c-0.3 0.4-0.6 0.7-0.9 1 -0.4 0.3-0.7 0.5-1.2 0.7 -0.4 0.2-0.9 0.2-1.3 0.2 -0.7 0-1.3-0.2-1.8-0.5 -0.5-0.3-0.9-0.7-1.2-1.2 -0.3-0.5-0.5-1-0.7-1.6 -0.1-0.6-0.2-1.2-0.2-1.9 0-0.9 0.1-1.7 0.4-2.4 0.3-0.7 0.6-1.4 1.1-1.9 0.5-0.6 1.1-1 1.8-1.3 0.7-0.3 1.5-0.5 2.3-0.5 0.7 0 1.4 0.1 2.1 0.3 0.6 0.2 1.3 0.5 1.8 0.9 0.6 0.4 1.1 0.8 1.6 1.3s0.9 1 1.4 1.5l-0.2 1.7c-0.2-0.3-0.5-0.6-0.9-0.9 -0.4-0.3-0.7-0.6-1.1-1 -0.4-0.3-0.8-0.6-1.2-0.9 -0.4-0.3-0.7-0.5-1-0.7L544.9 19.8zM541.2 26.3c0.3 0 0.5-0.1 0.8-0.4 0.2-0.3 0.5-0.6 0.7-1 0.2-0.4 0.3-0.9 0.5-1.4 0.1-0.5 0.2-1 0.2-1.4 0-0.4 0-0.7-0.1-1 -0.1-0.3-0.1-0.6-0.2-0.8 -0.1-0.2-0.2-0.4-0.4-0.5 -0.2-0.1-0.4-0.2-0.6-0.2 -0.3 0-0.6 0.1-0.9 0.4 -0.3 0.3-0.5 0.6-0.8 1 -0.2 0.4-0.4 0.9-0.5 1.4 -0.1 0.5-0.2 1-0.2 1.5 0 0.3 0 0.6 0.1 0.9 0.1 0.3 0.2 0.5 0.3 0.8 0.1 0.2 0.3 0.4 0.5 0.5C540.7 26.2 540.9 26.3 541.2 26.3z"
+    />
+    <path
+      className="flagst1"
+      d="M552.9 26.5c0.8 0 1.5-0.1 2.2-0.2 0.6-0.2 1.2-0.4 1.8-0.6 0.5-0.3 1-0.6 1.4-0.9 0.4-0.4 0.8-0.7 1.2-1.1l0.1 2.7c-0.2 0.3-0.6 0.6-0.9 0.9 -0.4 0.3-0.7 0.5-1.2 0.8 -0.4 0.2-0.9 0.4-1.3 0.5 -0.5 0.1-0.9 0.2-1.4 0.2 -0.1 0-0.1 0-0.2 0 -0.1 0-0.2 0-0.4-0.1 -0.2 0-0.5-0.1-0.9-0.2 -0.4-0.1-0.9-0.1-1.5-0.2v0.2c0 1.4-0.1 2.7-0.4 3.9 -0.2 1.2-0.6 2.2-1 3 -0.4 0.8-0.9 1.5-1.4 1.9 -0.5 0.4-1.1 0.7-1.6 0.7 -0.4 0-0.8-0.1-1.1-0.3 -0.3-0.2-0.5-0.5-0.7-0.9 -0.2-0.4-0.3-0.8-0.4-1.3 -0.1-0.5-0.1-1-0.1-1.6 0-0.3 0.1-0.7 0.2-1.1 0.1-0.4 0.2-0.8 0.4-1.3 0.2-0.4 0.4-0.9 0.6-1.3 0.2-0.5 0.4-0.9 0.7-1.3 0.5-1 1.1-2 1.8-3.1 0.1-0.6 0.2-1.2 0.3-1.9 0.1-0.6 0.2-1.3 0.3-1.9 0.1-0.6 0.3-1.2 0.5-1.7 0.2-0.5 0.5-1 0.8-1.4 0.3-0.4 0.7-0.7 1.2-0.9 0.5-0.2 1-0.4 1.6-0.4 0.5 0 0.9 0.1 1.3 0.2 0.4 0.1 0.8 0.3 1.1 0.6 0.3 0.3 0.6 0.7 0.7 1.1 0.2 0.5 0.3 1 0.3 1.7 0 0.5-0.1 1-0.2 1.5 -0.2 0.5-0.4 1-0.7 1.4 -0.3 0.5-0.8 0.9-1.3 1.3 -0.5 0.4-1.1 0.7-1.8 1.1L552.9 26.5zM547.4 33.7c0 0.1 0 0.2 0 0.3 0 0.1 0.1 0.3 0.1 0.4 0.1 0.1 0.1 0.3 0.2 0.4 0.1 0.1 0.2 0.2 0.4 0.2 0.1 0 0.2-0.1 0.3-0.2 0.1-0.1 0.2-0.3 0.2-0.6s0.1-0.5 0.1-0.8c0-0.3 0-0.6 0.1-0.9 0-0.3 0-0.6 0-0.9 0-0.3 0-0.6 0-0.8 0-0.4 0-0.8 0.1-1.1 0-0.3 0-0.5 0-0.7 0-0.2 0-0.4 0-0.6 -0.1 0.3-0.3 0.6-0.4 0.9 -0.1 0.3-0.3 0.7-0.4 1.1 -0.1 0.4-0.2 0.7-0.3 1.1 -0.1 0.4-0.2 0.7-0.3 1 -0.1 0.3-0.1 0.6-0.2 0.8C547.4 33.5 547.4 33.6 547.4 33.7zM551.4 25.3c0.5-0.2 0.9-0.5 1.2-0.8 0.3-0.3 0.6-0.6 0.9-0.9 0.2-0.3 0.4-0.7 0.6-1 0.2-0.4 0.3-0.7 0.4-1.1 0-0.2 0.1-0.4 0.1-0.6 0-0.2 0-0.4-0.1-0.7 -0.1-0.2-0.2-0.4-0.3-0.5 -0.1-0.1-0.3-0.2-0.6-0.2 -0.3 0-0.5 0.1-0.8 0.2 -0.2 0.2-0.4 0.4-0.6 0.7 -0.2 0.3-0.3 0.6-0.4 1 -0.1 0.4-0.2 0.8-0.3 1.2 -0.1 0.4-0.1 0.9-0.1 1.3C551.5 24.4 551.4 24.8 551.4 25.3z"
+    />
+    <path
+      className="flagst1"
+      d="M434.9 23.5c0.1 0 0.2 0 0.3 0 0.1 0 0.2 0.1 0.3 0.1 0.1 0 0.2 0.1 0.2 0.2 0.1 0.1 0.1 0.2 0.1 0.3 0 0.1 0 0.2 0 0.4 0 0.1-0.1 0.2-0.2 0.3 -0.1 0.1-0.2 0.2-0.3 0.2 -0.1 0.1-0.3 0.1-0.6 0.1H431c-0.2 0-0.3 0-0.4-0.1 -0.1-0.1-0.2-0.1-0.3-0.2 -0.1-0.1-0.1-0.2-0.1-0.3 0-0.1 0-0.2 0-0.3 0-0.1 0-0.2 0-0.2 0-0.1 0.1-0.2 0.1-0.2 0.1-0.1 0.1-0.1 0.3-0.2 0.1 0 0.3-0.1 0.5-0.1H434.9z"
+    />
+  </svg>
+);
+
+export default Flag;
