+++ client/components/Sto/index.js
@@ -1,0 +1,31 @@
+import React from "react";
+
+const Sto = () => (
+  <svg
+    width={25}
+    height={25}
+    viewBox="0 0 120 120"
+    style={{ enableBackground: "new 0 0 120 120" }}
+  >
+    <style jsx>
+      {`
+        .stost0 {
+          fill: #bcbbbb;
+        }
+        .stost1 {
+          fill: #f48023;
+        }
+      `}
+    </style>
+    <polygon
+      className="stost0"
+      points="84.4,93.8 84.4,70.6 92.1,70.6 92.1,101.5 22.6,101.5 22.6,70.6 30.3,70.6 30.3,93.8 "
+    />
+    <path
+      className="stost1"
+      d="M38.8,68.4l37.8,7.9l1.6-7.6l-37.8-7.9L38.8,68.4z M43.8,50.4l35,16.3l3.2-7l-35-16.4L43.8,50.4z M53.5,33.2 l29.7,24.7l4.9-5.9L58.4,27.3L53.5,33.2z M72.7,14.9l-6.2,4.6l23,31l6.2-4.6L72.7,14.9z M38,86h38.6v-7.7H38V86z"
+    />
+  </svg>
+);
+
+export default Sto;
